% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{annotation_custom2}
\alias{annotation_custom2}
\title{This is a special geom intended for use as static annotations derived from
ggplot2::annotation_custom() matching a specific panel on a faceted ggplot.}
\usage{
annotation_custom2(
  grob,
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf,
  data
)
}
\arguments{
\item{grob}{A \code{grob} to display.}

\item{xmin, xmax}{x location (in data coordinates) giving horizontal location
of raster.}

\item{ymin, ymax}{y location (in data coordinates) giving vertical location
of raster.}

\item{data}{A subset of a \code{data.table} matching the panel where the
grob annotation should be displayed.}
}
\description{
This is a special geom intended for use as static annotations derived from
ggplot2::annotation_custom() matching a specific panel on a faceted ggplot.
}
\keyword{internal}
