% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{basic.sidebar}
\alias{basic.sidebar}
\title{Draws a ggplot2 of a basic sidebar.}
\usage{
basic.sidebar(data, palette)
}
\arguments{
\item{data}{A \code{data.frame} with the column names 'Samples','.id' and
'Groups' in this order.}

\item{palette}{A \code{character} vector containing R colors like a palette.}
}
\value{
A \code{gg} object of the basic sidebar (a 'geom_tile()').
}
\description{
Draws a ggplot2 of a basic sidebar.
}
\author{
Yoann Pageaud.
}
