% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{build.col_table}
\alias{build.col_table}
\title{Builds color tables for legends.}
\usage{
build.col_table(annot.grps, annot.pal)
}
\arguments{
\item{annot.grps}{A \code{type} parameter description.}
}
\value{
A \code{data.table} list of categories matching their colors for each
        legend.
}
\description{
Builds color tables for legends.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
