% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{build.ggvolcano}
\alias{build.ggvolcano}
\title{Core function that builds a volcano plot using ggplot2.}
\usage{
build.ggvolcano(
  data,
  data.type,
  label.cutoff = 0,
  p.cutoff = 0.01,
  y.cutoff = l2fc.cutoff,
  title.y.cutoff = "Y-Axis cutoff"
)
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Core function that builds a volcano plot using ggplot2.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
