% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plots.R
\name{build_asso_plot}
\alias{build_asso_plot}
\title{Builder for the association plot using ggplot2.}
\usage{
build_asso_plot(asso_res, dataset.name = "dataset", PC.origin = NULL)
}
\arguments{
\item{asso_res}{A \code{data.table} generated by
\link{test_asso_annot_pc} containing association tests'
results between a table of annotation and a set of
principal components.}

\item{dataset.name}{A \code{character} string specifying the name you want to
give to your dataset, that will be included in the plot
title (Default: dataset.name = "dataset").}

\item{PC.origin}{A \code{character} string specifying what kind of data
the prcomp object has been computed from, especially if
it has been computed on a subset of the initial dataset
(e.g. 'PCs from ...') it will be appended to the plot
title (Default: PC.origin = NULL).}
}
\value{
A \code{gg} plot summarizing results from associations tests between
        dataset's annotations and principal components.
}
\description{
Builder for the association plot using ggplot2.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
