% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{chk.breaks.labels}
\alias{chk.breaks.labels}
\title{Checks if n.breaks and labels match in a ScaleContinuous object.}
\usage{
chk.breaks.labels(ScaleContinuous_obj)
}
\arguments{
\item{ScaleContinuous_obj}{A \code{ScaleContinuous} object usually generated
with \link[ggplot2]{scale_fill_gradient},
\link[ggplot2]{scale_fill_gradient2} or
\link[ggplot2]{scale_fill_gradientn}.}
}
\description{
Checks if n.breaks and labels match in a ScaleContinuous object.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
