% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{chk.corr.lbl.cutoff}
\alias{chk.corr.lbl.cutoff}
\title{Checks the correlation cut-off value(s) given and computes the negative and
positive correlation cut-offs.}
\usage{
chk.corr.lbl.cutoff(corr.label.cutoff)
}
\arguments{
\item{corr.label.cutoff}{A \code{numeric} vector of 1 or 2 values, between -1
and 1 to be used as the minimum cut-off on
positive and negative correlation values.
\itemize{
 \item{If 1 value is given, it will be define as the
 minimum cut-off on absolute correlation values
 (positive and negative ones).}
 \item{If 2 values are given, the smalest one will
 be used as a maximum cut-off on negative
 correlation values. the biggest one will be used as
 a minimum cut-off on positive correlation values.
 The smalest value must be comprised between -1 and
 0. The biggest value must be comprised between 0
 and 1.}
}}
}
\value{
A \code{numeric} vector of length 2 containing the negative cut-off
        and the positive cut-off for correlation values.
}
\description{
Checks the correlation cut-off value(s) given and computes the negative and
positive correlation cut-offs.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
