% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{chk.param}
\alias{chk.param}
\title{Checks parameters for tests functions}
\usage{
chk.param(data, data.type, label.cutoff)
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Checks parameters for tests functions
}
\references{

}
\author{
Yoann Pageaud.
}
\keyword{internal}
