% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdensity_map.R
\name{colDensity}
\alias{colDensity}
\title{Computes density by column on a matrix.}
\usage{
colDensity(
  x,
  na.rm = FALSE,
  from = min(x, na.rm = TRUE),
  to = max(x, na.rm = TRUE),
  ncores = 1
)
}
\arguments{
\item{x}{A \code{matrix}.}

\item{na.rm}{A \code{logical}. Should missing values (including NaN) be
omitted from the calculations?}

\item{from}{A \code{numeric} to specify the minimum value from which the
kernel estimation starts
(Default: from = min(x, na.rm = TRUE)).}

\item{to}{A \code{numeric} to specify the maximum value to which the
kernel estimation ends (Default: to = max(x, na.rm = TRUE)).}

\item{ncores}{An \code{integer} to specify the number of cores/threads to be
used to parallel-compute densities on matrix columns.}
}
\value{
A \code{list} of data.tables containing the density data, one
        data.table per column.
}
\description{
Computes density by column on a matrix.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
