% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbipca.R
\name{cross.biplot}
\alias{cross.biplot}
\title{Computes and draws biplots for multiple principal components at once.}
\usage{
cross.biplot(
  prcomp.res,
  data,
  PCs = c(1, 2, 3),
  scale = 1,
  point.size = 1,
  color.data = NULL,
  fill.data = NULL,
  shape.data = NULL,
  loadings = FALSE,
  loadings.col = "red",
  top.load.by.quad = NULL
)
}
\arguments{
\item{prcomp.res}{A PCA result of classes \code{prcomp} or
\code{irlba_prcomp} resulting from stats::prcomp()
or irlba::prcomp_irlba().}

\item{data}{A \code{data.table} containing annotations matching
data in 'prcomp.res'.}

\item{PCs}{An \code{integer} vector matching principal
components to be used to generate the cross-biplot.}

\item{scale}{A \code{double} scaling parameter, disabled by 0.}

\item{point.size}{A \code{double} specifying the size of points.}

\item{color.data}{A \code{character} specifying the column name in
'data' to be used to map colors to points.}

\item{shape.data}{A \code{character} specifying the column name in
'data' to be used to map shapes to points.}

\item{loadings}{A \code{logical} specifying whether the loadings
should be displayed (TRUE) or not (FALSE).}

\item{loadings.col}{A \code{character} specifying a color to be used
for loadings.}

\item{top.load.by.quad}{An \code{integer} specifying the top n most
important loadings to be displayed in the four
quadrants of the biplot graph (by quadrants). This
parameters allows to display only the most
important loadings, and to hide the less important
ones, to improve visibility when there is too many
of them.}
}
\value{
A \code{gg} plot displaying all combinations of biplots for the
        selected PCs.
}
\description{
Computes and draws biplots for multiple principal components at once.
}
\examples{
#Get PCA results
pca.res <- prcomp(iris[,-5])
#Draw the simplest cross-biplot:
cross.biplot(prcomp.res = pca.res, data = iris)
#Draw a biplot of PC1 to PC4:
cross.biplot(prcomp.res = pca.res, data = iris, PCs = c(1:4))
#Change data scale:
cross.biplot(prcomp.res = pca.res, data = iris, scale = 2)
#Change points size:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2)
#Color points following Species:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             color.data = "Species")
#Change points shape following Species:
# (could use any other categorical / ordinal value if others available)
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             color.data = "Species", shape.data = "Species")
#Map custom colors:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             color.data = "Species", shape.data = "Species") +
  scale_color_manual(values = c("green", "red", "orange"))
#Map custom point shapes:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             color.data = "Species", shape.data = "Species") +
  scale_color_manual(values = c("green", "red", "orange")) +
  scale_shape_manual(values = c(83, 8, 25))
#Map custom filling colors:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             fill.data = "Species", shape.data = "Species") +
  scale_fill_manual(values = c("green", "royalblue", "orange")) +
  scale_shape_manual(values = c(21, 22, 23)) # Needs values between 21 and 25
#Enable custom filling colors and fix point shapes:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             fill.data = "Species", shape.data = "Species") +
  scale_fill_manual(values = c("green", "royalblue", "orange")) +
  scale_shape_manual(values = rep(21, 3)) # Maps shape N°21 to 3 categories
#Map filling colors, outline colors, and point shapes all together:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             fill.data = "Species", shape.data = "Species",
             color.data = "Species") +
  scale_fill_manual(values = c("green", "royalblue", "orange")) +
  scale_color_manual(values = c("black", "red", "blue")) +
  scale_shape_manual(values = c(21, 22, 23)) # Needs values between 21 and 25
#Show loadings:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             color.data = "Species", shape.data = "Species",
             loadings = TRUE) +
  scale_color_manual(values = c("green", "red", "orange")) +
  scale_shape_manual(values = c(83, 8, 25))
#Change loadings color:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             color.data = "Species", shape.data = "Species", loadings = TRUE,
             loadings.col = "purple") +
  scale_color_manual(values = c("green", "red", "orange")) +
  scale_shape_manual(values = c(83, 8, 25))
#Display the top 1 loading in each quadrant:
cross.biplot(prcomp.res = pca.res, data = iris, point.size = 2,
             color.data = "Species", shape.data = "Species", loadings = TRUE,
             loadings.col = "purple", top.load.by.quad = 1) +
  scale_color_manual(values = c("green", "red", "orange")) +
  scale_shape_manual(values = c(83, 8, 25))
}
\author{
Yoann Pageaud.
}
