% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva.R
\name{EVA}
\alias{EVA}
\title{Computes eigenvectors, PC scores and correlations from a correlation test.}
\usage{
EVA(
  data,
  use = "pairwise",
  method = "pearson",
  adjust = "none",
  var.min = 0.01,
  groups = as.character(seq(ncol(data))),
  colors = grDevices::rainbow(n = ncol(data)),
  label = TRUE
)
}
\arguments{
\item{data}{A \code{matrix} or a \code{data.frame} containing
variables by columns and values to be used for the
correlation test by rows.}

\item{use}{A \code{character} to specify how to handle missing values
when calculating a correlation. Possible values are
'pairwise' and 'complete'. 'pairwise' is the default value
and will do pairwise deletion of cases. 'complete' will
select just complete cases.}

\item{method}{The correlation method to use as a \code{character}
matching one of these: 'pearson','spearman','kendall'.}

\item{adjust}{A \code{character} specifying what adjustment for multiple
tests should be used. Possible values are: "holm",
"hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" and
"none".}

\item{var.min}{A \code{double} setting the minimum variance accountable
for an eigenvector to be considered in the plots
generated.}

\item{groups}{A \code{character} vector of groups to which variables
belong for eigenvector annotations. The length of this
vector has to match the number/variables of columns in the
data.}

\item{colors}{A \code{character} vector of colors for the eigenvectors.
The length of this vector has to match the number of
different groups existing.}
}
\value{
A \code{list} containing the principal components correlations,
        a list of the Eigenvector Plots generated and principal components
        scores of the scaled data.
}
\description{
Computes eigenvectors, PC scores and correlations from a correlation test.
}
\author{
Yoann Pageaud.
}
