% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_histogram.R
\name{fancy.hist}
\alias{fancy.hist}
\title{Computes in parallel and plot an histogram using ggplot2 from a given vector
of values.}
\usage{
fancy.hist(
  x,
  xmax,
  nbreaks = 10,
  ngrad = 10,
  round.grad = 1,
  ncores = 1,
  xlab = "values",
  ylab = "Frequency",
  bin.col = "#0570b0"
)
}
\arguments{
\item{x}{A \code{numeric} vector to be used for plotting the
histogram.}

\item{xmax}{A \code{numeric} specifying the maximum limit on the x axis
to display values. All values above this limit will be
agregated on the last histogram bin.}

\item{nbreaks}{An \code{integer} specifying the number of delimitations to
be use for histogram bins\cr(Default: nbreaks = 10).}

\item{ngrad}{An \code{integer} specifying the number of graduations to
display on the X-axis of the plot\cr(Default: ngrad = 10).}

\item{round.grad}{An \code{integer} specifying the number significant digits
to be considered when calculating graduations\cr
(Default: round.grad = 1).}

\item{ncores}{An \code{integer} to specifying the number of cores to use
when parallel-running the computation of the histogram\cr
(Default: ncores = 1).}

\item{xlab}{A \code{character} to be used as X-axis label\cr
(Default: xlab = 'values').}

\item{ylab}{A \code{character} to be used as Y-axis label\cr
(Default: ylab = 'Frequency').}

\item{bin.col}{A \code{character} matching a R color code to be use to
fill the histogram bins.}
}
\value{
A \code{gg} plot of an histogram.
}
\description{
Computes in parallel and plot an histogram using ggplot2 from a given vector
of values.
}
\author{
Yoann Pageaud.
}
