% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfusion.R
\name{fix.corrMatOrder.alphabet}
\alias{fix.corrMatOrder.alphabet}
\title{Fixes a bug related to corrMatOrder when order parameter is set to
'alphabet'.}
\usage{
fix.corrMatOrder.alphabet(cor.order, str)
}
\arguments{
\item{cor.order}{A \code{character} vector of interest.}

\item{str}{A \code{character} vector from which to get elements order.}
}
\value{
A \code{type} object returned description.
}
\description{
Fixes a bug related to corrMatOrder when order parameter is set to
'alphabet'.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
