% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_plot.R
\name{fused.plot}
\alias{fused.plot}
\title{Creates a plot summarizing results from 2 different pairwise comparisons.}
\usage{
fused.plot(
  data,
  ncores,
  upper.comp,
  upper.value,
  lower.comp,
  lower.value,
  na.rm = "pairwise",
  order.method,
  order.select,
  hclust.method = "complete",
  p.adjust,
  annot.grps = list(Groups = seq(ncol(data))),
  annot.pal = rainbow(n = ncol(data)),
  annot.pos = "top",
  annot.size = 0,
  annot.lgd.merge = FALSE,
  annot.split = TRUE,
  dendro.pos = "none",
  dendro.size = 0,
  grid.col = "grey",
  grid.thickness = 0.5,
  axis.title = element_blank(),
  axis.title.x = element_blank(),
  axis.title.y = element_blank(),
  axis.text = element_text(size = 12),
  axis.text.x = element_text(angle = 90, hjust = 0, vjust = 0.5),
  axis.text.y = element_blank(),
  axis.ticks = element_line(color = "black"),
  set.x.title = NULL,
  set.y.title = NULL,
  set.lgd1.title = NULL,
  set.lgd2.title = NULL,
  diag.col = "white",
  lgd.pal1 = NULL,
  lgd.pal2 = NULL,
  lgd.title = element_blank(),
  lgd.title1 = element_blank(),
  lgd.title2 = element_blank(),
  lgd.text = element_text(size = 12),
  lgd.text1 = element_blank(),
  lgd.text2 = element_blank(),
  lgd.breaks = NULL,
  lgd.breaks1 = NULL,
  lgd.breaks2 = NULL,
  lgd.labels = NULL,
  lgd.labels1 = NULL,
  lgd.labels2 = NULL,
  lgd.round = NULL,
  lgd.round1 = 2,
  lgd.round2 = 2,
  lgd.limits = NULL,
  lgd.limits1 = NULL,
  lgd.limits2 = NULL,
  lgd.ticks = TRUE,
  lgd.ticks1 = NULL,
  lgd.ticks2 = NULL,
  lgd.ticks.linewidth = 2,
  lgd.ticks.linewidth1 = NULL,
  lgd.ticks.linewidth2 = NULL,
  lgd.nbin = NULL,
  lgd.nbin1 = NULL,
  lgd.nbin2 = NULL,
  lgd.height1 = 26,
  lgd.height2 = 1,
  lgd.width1 = 1,
  lgd.width2 = 30,
  lgd.frame.col = "grey",
  lgd.frame.linewidth = 1.5,
  lgd.frame.linewidth1 = NULL,
  lgd.frame.linewidth2 = NULL,
  raster = FALSE,
  raster1 = NULL,
  raster2 = NULL,
  add.ggplot.arg = NULL
)
}
\arguments{
\item{data}{A \code{matrix} or \code{dataframe}.}

\item{ncores}{An \code{integer} to specify the number of
cores/threads to be used to parallel-run tests.}

\item{upper.comp}{The comparison for which results will be
displayed in the upper triangle of the plot as a
\code{character} matching one of these:
'pearson','spearman','kendall'.}

\item{upper.value}{A \code{character} matching the type of value to
                            display in the upper triangle of the plot
                            resulting from the comparison.\cr The possible
                            value types are:
\itemize{
 \item{'r' - the correlation values.}
 \item{'n' - the number of cases used for the comparisons.}
 \item{'stat' - the values of the comparison statistic.}
 \item{'p' - the P-values of the comparison.}
 \item{'se' - the standard errors of the comparison.}
}}

\item{lower.comp}{The comparison for which results will be
displayed in the lower triangle of the plot as a
\code{character} matching one of these:
'pearson','spearman','kendall'.}

\item{lower.value}{See 'upper.value' help.}

\item{na.rm}{A \code{character} to specify how to handle
missing values when calculating a correlation.
Possible types are 'pairwise' and 'complete'.
'pairwise' is the default value and will do
pairwise deletion of cases. 'complete' will
select just complete cases.}

\item{order.method}{A \code{character} specifying the ordering
                             method to apply.\cr Possible ordering methods are:
\itemize{
 \item{'AOE' - angular order of the eigenvectors.}
 \item{'FPC' - first principal component order.}
 \item{'hclust' - hierarchical clustering order.}
 \item{'alphabet' - alphabetical order}
 \item{'default'}
}}

\item{order.select}{A \code{character} specifying comparison results
to use for the clustering.\cr Possible values
are 'upper' or 'lower'.}

\item{hclust.method}{A \code{character} specifying the method to use
for the hierarchical clustering if the ordering
method is 'hclust'.\cr Possible methods are:
'ward.D','ward.D2', 'single', 'complete',
'average' (= UPGMA), 'mcquitty' (= WPGMA),
median' (= WPGMC) or 'centroid' (= UPGMC).}

\item{p.adjust}{A \code{character} specifying what adjustment
for multiple tests should be used.\cr Possible
values are: "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none".}

\item{annot.grps}{A \code{list} of vectors of groups to which
variables belongs for the annotation sidebars.
Vectors' lengths have to match the number of
variables.}

\item{annot.pal}{A \code{vector} or a list of vectors containing
colors as characters for the annotation
sidebars. The length of vectors has to match the
number of levels of vectors listed in
'annot.grps'. If a list is provided, its length
must match the length of the list provided to
'annot.grps'.}

\item{annot.pos}{A \code{character} specifying the position of
the annotation sidebar.\cr Possible values are:
'top', 'left' or 'both'.}

\item{annot.size}{An \code{integer} to increase or decrease the
size of the annotation side bar.}

\item{dendro.pos}{A \code{character} specifying the position of
dendrograms if the selected order is 'hclust'.
\cr Possible values are: 'top','left','none'.}

\item{dendro.size}{An \code{integer} to increase or decrease the
size of the dendrogram.}

\item{grid.col}{A \code{character} specifying the color of the
grid.}

\item{grid.thickness}{A \code{double} value for the thickness of the
grid.}

\item{axis.title}{An \code{element_text} object to setup axes
titles.}

\item{axis.title.x}{An \code{element_text} object to setup X axis
title.}

\item{axis.title.y}{An \code{element_text} object to setup Y axis
title.}

\item{axis.text}{An \code{element_text} object to setup axes
text.}

\item{axis.text.x}{An \code{element_text} object to setup X axis
text.}

\item{axis.text.y}{An \code{element_text} object to setup Y axis
text.}

\item{axis.ticks}{An \code{element_line} object to setup the ticks
of the plot.}

\item{set.x.title}{A \code{character}to be used as the title for
the X axis.}

\item{set.y.title}{A \code{character}to be used as the title for
the Y axis.}

\item{set.lgd1.title}{A \code{character}to be used as the title of the
legend.}

\item{set.lgd2.title}{A \code{character}to be used as the title of the
legend.}

\item{diag.col}{A \code{character} defining the color of cells
with of the empty diagonal.}

\item{lgd.pal1}{A \code{character} vector of colors to use for
the palette of the upper triangle.}

\item{lgd.pal2}{A \code{character} vector of colors to use for
the palette of the lower triangle.}

\item{lgd.title}{An \code{element_text} object to setup both
legend titles.}

\item{lgd.title1}{An \code{element_text} object to setup the
legend title of the upper triangle.}

\item{lgd.title2}{An \code{element_text} object to setup the
legend title of the lower triangle.}

\item{lgd.text}{An \code{element_text} object to setup the text
of both legends.}

\item{lgd.text1}{An \code{element_text} object to setup the text
of the upper triangle legend.}

\item{lgd.text2}{An \code{element_text} object to setup the text
of the lower triangle legend.}

\item{lgd.breaks}{A \code{double} vector defining the graduation
position along the possible values.}

\item{lgd.breaks1}{A \code{double} vector defining the graduation
position along the possible values of the upper
triangle.}

\item{lgd.breaks2}{A \code{double} vector defining the graduation
position along the possible values of the lower
triangle.}

\item{lgd.labels}{A \code{double} vector of values to map to
legends breaks. Will be displayed on the plot.}

\item{lgd.labels1}{A \code{double} vector of values to map to the
upper triangle legend breaks. Will be displayed
on the plot.}

\item{lgd.labels2}{A \code{double} vector of values to map to the
lower triangle legend breaks. Will be displayed
on the plot.}

\item{lgd.round}{An \code{integer} indicating the number of
decimal places to be used for the default
legends labels.}

\item{lgd.round1}{An \code{integer} indicating the number of
decimal places to be used for the default
upper triangle legend labels.}

\item{lgd.round2}{An \code{integer} indicating the number of
decimal places to be used for the default
lower triangle legend labels.}

\item{lgd.limits}{A \code{double} vector of length 2, giving the
upper and lower limits for mapping breaks and
colors to both triangle legends.}

\item{lgd.limits1}{A \code{double} vector of length 2, giving the
upper and lower limits for mapping breaks and
colors to the upper triangle legend.}

\item{lgd.limits2}{A \code{double} vector of length 2, giving the
upper and lower limits for mapping breaks and
colors to the lower triangle legend.}

\item{lgd.ticks.linewidth}{A \code{double} value to specify the thickness
of legends ticks.}

\item{lgd.ticks.linewidth1}{A \code{double} value to specify the thickness
of the upper triangle associated legend ticks.}

\item{lgd.ticks.linewidth2}{A \code{double} value to specify the thickness
of the lower triangle associated legend ticks.}

\item{lgd.nbin}{An \code{integer} specifying the number of bins
for drawing both colorbars. A smoother
colorbar results from a larger value.}

\item{lgd.nbin1}{An \code{integer} specifying the number of bins
for drawing the upper triangle associated
colorbar. A smoother colorbar results from a
larger value.}

\item{lgd.nbin2}{An \code{integer} specifying the number of bins
for drawing the lower triangle associated
colorbar. A smoother colorbar results from a
larger value.}

\item{lgd.height1}{A \code{double} specifying the height of the
upper triangle associated colorbar.}

\item{lgd.height2}{A \code{double} specifying the height of the
lower triangle associated colorbar.}

\item{lgd.width1}{A \code{double} specifying the width of the
upper triangle associated colorbar.}

\item{lgd.width2}{A \code{double} specifying the height of the
lower triangle associated colorbar.}

\item{lgd.frame.col}{A \code{character} defining the color of the
frame of legends.}

\item{lgd.frame.linewidth}{A \code{double} defining the thickness of the
frame of both legends.}

\item{lgd.frame.linewidth1}{A \code{double} defining the thickness of the
frame of upper triangle associated legends.}

\item{lgd.frame.linewidth2}{A \code{double} defining the thickness of the
frame of lower triangle associated legends.}

\item{raster}{A \code{logical} to specify whether or not
colors in both legends should be rasterized.}

\item{raster1}{A \code{logical} to specify whether or not
colors in the upper triangle associated legend
should be rasterized.}

\item{raster2}{A \code{logical} to specify whether or not
colors in the lower triangle associated legend
should be rasterized.}

\item{add.ggplot.args}{Ultimate ggplot2 additionnal components that you
want to pass to the plot (Use only if you are an
ggplot2 advanced user).}
}
\value{
A \code{gtable} object plotted automatically and a \code{list} of
       results from the 2 comparisons as 2 matrices, with the same gtable
       object:
       \itemize{
        \item{'upper.res' - the matrix result of pairwise comparisons
        displayed in the upper triangle.}
        \item{'lower.res' - the matrix result of pairwise comparisons
        displayed in the lower triangle.}
        \item{'fused.plot' - the plot.}
       }
}
\description{
Creates a plot summarizing results from 2 different pairwise comparisons.
}
\author{
Yoann Pageaud.
}
