% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{get.len.legends}
\alias{get.len.legends}
\title{Gets legend lengths.}
\usage{
get.len.legends(col_table)
}
\arguments{
\item{col_table}{A \code{data.table} list of categories with their matching
color palettes for legends.}
}
\value{
An \code{integer} list of sizes, one per legend to be displayed.
}
\description{
Gets legend lengths.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
