% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{get.lgd}
\alias{get.lgd}
\title{Extracts legend from a ggplot2 object.}
\usage{
get.lgd(gg2.obj)
}
\arguments{
\item{gg2.obj}{A \code{gg} object with legends.}
}
\value{
A \code{gg} object only containing the legends of the plot.
}
\description{
Extracts legend from a ggplot2 object.
}
\references{
\href{https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}{Share a legend between two ggplot2 graphs - Mara Averick}
}
\author{
Yoann Pageaud.
}
\keyword{internal}
