% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stat_tests.R
\name{get_ks_stat}
\alias{get_ks_stat}
\title{Extracts Kolmogorov-Smirnov test results and return them as a matrix.}
\usage{
get_ks_stat(table_combinations, dt_ks_tests, statistic)
}
\arguments{
\item{dt_ks_tests}{A \code{data.table} containing raw results from multiple
Kolmogorov-Smirnov tests.}

\item{statistic}{A \code{character} specifying the type of statistic to
retrieve from the test\cr
(Supported: statistic = c('n','stat','p')).}
}
\value{
A \code{matrix} containing the KS test's statistic values of
        interest.
}
\description{
Extracts Kolmogorov-Smirnov test results and return them as a matrix.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
