% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbipca.R
\name{get_loadings}
\alias{get_loadings}
\title{Checks classes of a prcomp object to extract loadings data.}
\usage{
get_loadings(prcomp.res, PCs)
}
\arguments{
\item{prcomp.res}{A PCA result of classes \code{prcomp} or
\code{irlba_prcomp} resulting from stats::prcomp() or
irlba::prcomp_irlba().}
}
\value{
A \code{data.table} containing data to plot loadings on a PCA biplot.
}
\description{
Checks classes of a prcomp object to extract loadings data.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
