% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbivar.R
\name{ggbivar}
\alias{ggbivar}
\title{Draws boxplots or violins from a variable values against ranges of a 2nd one.}
\usage{
ggbivar(
  data,
  violin = FALSE,
  cat.step = 10L,
  cat.max = 10L,
  fill = "deepskyblue3"
)
}
\arguments{
\item{data}{A two-columns \code{data.frame}:
\itemize{
 \item{column 1 contains values of the 1st variable. These
 values will be used for computing boxplots.}
 \item{column 2 contains values of the 2nd variable. These
 values will be used to define ranges, one range per
 boxplot.}
}}

\item{violin}{A \code{logical} to specify whether violins of categories
should be displayed with the boxplots (Warning: setting
violin = TRUE significantly increase the computing time.
Default: violin = FALSE).}

\item{cat.step}{A \code{numeric} to specify the size of ranges.
(Default: cat.step = 10L).}

\item{cat.max}{Maximum amount of ranges to create (Default: cat.max = 10L).
Values outside the last range are grouped into an extra bin
named "over <cat.step * cat.max>". In total the plot will
display 1 boxplot for each range, and one additional boxplot
for the values falling into the extra bin.}

\item{fill}{A \code{character} to specify a color to be used for filling
boxplot. The color will be used to derive light shades from
it. The given input color will match the color of the boxplot
computed on the highest number of values (in other words: for
the largest category). Consequently, all the shades generated
will be lighter than the original one. Thus, it is advised to
use a dark color (Default: fill = "deepskyblue3").}
}
\value{
A \code{gg} plot object displaying boxplots and/or violin plots.
}
\description{
Draws boxplots or violins from a variable values against ranges of a 2nd one.
}
\examples{
# Draw a bivar boxplot with a step of 1
ggbivar(data = iris[, c(1,3)], cat.step = 1)
# Draw a bivar violin plot with a step of 1
ggbivar(data = iris[, c(1,3)], cat.step = 1, violin = TRUE)
# Change the filling color of the bivar plot
ggbivar(data = iris[, c(1,3)], cat.step = 1, fill = "red")
# Aggregate values above 5
ggbivar(data = iris[, c(1,3)], cat.step = 1, fill = "red", cat.max = 5)
# Example of customization of axis titles using ggplot2 syntax
ggbivar(data = iris[, c(1,3)], cat.step = 1, fill = "red", cat.max = 5) +
  labs(x = "Petal length", y = "Sepal length")
}
\author{
Yoann Pageaud.
}
