% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcirclart.R
\name{ggcirclart}
\alias{ggcirclart}
\title{Circlizes ggplot2 objects.}
\usage{
ggcirclart(
  gg,
  facet_fun = "wrap",
  wrap_to = "x",
  wrap_nrow = NULL,
  wrap_ncol = NULL
)
}
\arguments{
\item{gg}{A \code{gg} plot object built with ggplot2.}

\item{facet_fun}{A \code{character} specifying the facet function to use for
the circlized plot, if the input gg object has facets. To
use facet_grid(): facet_fun = "grid"; To use facet_wrap():
facet_fun = 'wrap'.}

\item{wrap_to}{A \code{character} specifying to which axis the facet should
be assigned, when the input gg object uses facet_wrap() and
facet_fun = 'grid'. Supported values: wrap_to = 'x' or 'y'.}

\item{wrap_nrow}{An \code{integer} specifying on how many rows the facets
should be wrapped when facet_fun = 'wrap'.}

\item{wrap_ncol}{An \code{integer} specifying on how many columns the facets
should be wrapped when facet_fun = 'wrap'.}
}
\value{
A circlized \code{gg} object.
}
\description{
Circlizes ggplot2 objects.
}
\examples{
# Most basic circlizing on a histogram generated with BiocompR::gghist()
distrib <- rnorm(100, 5, 1)
original_hist <- BiocompR::gghist(x = distrib)
original_hist
ggcirclart(gg = original_hist)
# Circlizing on a faceted histogram generated with BiocompR::gghist() using
# facet_wrap()
faceted_hist <- BiocompR::gghist(
    x = distrib, facet = rep(x = c("A", "B", "C", "D"), each = 25))
faceted_hist
ggcirclart(gg = faceted_hist, facet_fun = "wrap")
# Wrap on only 1 row
ggcirclart(gg = faceted_hist, facet_fun = "wrap", wrap_nrow = 1)
# Wrap on 4 columns
ggcirclart(gg = faceted_hist, facet_fun = "wrap", wrap_ncol = 4)
# Circlizing on a faceted histogram generated with BiocompR::gghist() using
# facet_grid()
ggcirclart(gg = faceted_hist, facet_fun = "grid")
# Map wrapping variable to Y axis (instead of X axis, the default)
ggcirclart(gg = faceted_hist, facet_fun = "grid", wrap_to = "y")
# Customize your circlized ggplot2 object
ggcirclart(gg = faceted_hist, facet_fun = "wrap") +
    theme(
        plot.title = element_text(hjust = 0.5),
        axis.text = element_text(size = 11, color = "black"),
        axis.title = element_text(size = 13),
        panel.grid.major.y = element_line(color = "black", linewidth = 0.2),
        panel.border = element_rect(colour = "black", fill = NA),
        strip.background = element_rect(color = "black"),
        strip.text = element_text(size = 10.5),
        panel.spacing = unit(0, "lines"))
}
\author{
Yoann Pageaud.
}
