% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GG_Cluster_Heatmap.R
\name{ggclust.heatmap}
\alias{ggclust.heatmap}
\title{Create a clustered heatmap using GGplot.}
\usage{
ggclust.heatmap(data, regions.clusters, order.type = "Hclust",
  distance.method, select.clust = NULL, select.sample = NULL,
  groups = NULL, groups.order = NULL, sampling = F,
  sampling.size = 500, na.rm = F, title = "",
  palette = c("steelblue", "gray95", "darkorange"), na.col = "black",
  margins = c(0.5, 0, 0, 0))
}
\arguments{
\item{data}{A \code{matrix} or a \code{data.frame} with samples
by columns and regions by rows. Row names have to be
of the format "chr:start-end" and matching at least
some regions clustered.}

\item{regions.clusters}{A \code{data.frame} of BED format with the cluster
IDs (chr, start, end, clusterID).}

\item{order.type}{A \code{character} specifying how regions should be
ordered within a cluster. 2 possible methods:
SD - ordered by increasing order of standard
deviation
Hclust - ordered following a hierarchical clustering
method.}

\item{distance.method}{A \code{character} specifying the distance
calculation method if regions should be ordered
following a hierarchical clustering method.
Distance methods for continuous input variables:
bhjattacharyya, bray, canberra, chord, divergence,
dtw, euclidean, fJaccard, geodesic, hellinger,
kullback, mahalanobis, manhattan, maximum, minkowski,
podani, soergel, wave, whittaker.
Distance methods for binary input variables:
binary, braun-blanquet, cosine, dice, fager, faith,
hamman, hamming, kulczynski1, kulczynski2, michael,
mountford, mozley, ochiai, phi, russel,
simple matching,  simpson, stiles, tanimoto, yule,
yule2.}

\item{select.clust}{A \code{character} vector specifying the clusters to
keep for the plot.}

\item{select.sample}{A \code{character} vector specifying the samples to
keep for the plot (must match the columns names of
the matrix).}

\item{groups}{A \code{character} vector specifying the name of the
groups for samples. The length of the vector must
match the number of columns in the matrix.}

\item{groups.order}{A \code{character} vector specifying the names of the
groups (must be unique) in the wanted order.}

\item{sampling}{A \code{logical} to specify if a random sampling
should be applied on each cluster.}

\item{sampling.size}{An \code{integer} specifying the number of regions to
randomly select from each cluster if sampling = TRUE.
If one of the clusters contains less regions than the
amount specified, all regions of the cluster will be
selected.}

\item{na.rm}{A \code{logical} to specify if NAs should be removed
or not from the data. If order.type = "Hclust", NAs
are removed anyway.}

\item{title}{A \code{character} to be used as the plot title.}

\item{palette}{A \code{character} vector to be used as a color
palette to plot the data.}

\item{na.col}{A \code{character} matching a color to used for NAs
on the plot.}

\item{margins}{A \code{double} vector of length 4 to give the
thickness of each margin of the plot.}
}
\value{
A \code{gg} plot.
}
\description{
Create a clustered heatmap using GGplot.
}
\author{
Yoann Pageaud.
}
