% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{ggclust_layout}
\alias{ggclust_layout}
\title{Proof of concept for making a multi-panel clustered ggplot2 heatmap.}
\usage{
ggclust_layout(ls_col_ggdend, ls_ggannot, ls_row_ggdend, ls_ggheatmap)
}
\arguments{
\item{ls_col_ggdend}{A \code{list} of ggplot2 dendrograms on matrix colums.}

\item{ls_ggannot}{A \code{list} of ggplot2 geom_tile() serving as color
annotation sidebars.}

\item{ls_row_ggdend}{A \code{list} of ggplot2 dendrograms on matrix rows.}

\item{ls_ggheatmap}{A \code{list} of ggplot2 geom_tile() serving as
heatmaps.}
}
\value{
A \code{type} object returned description.
}
\description{
Proof of concept for making a multi-panel clustered ggplot2 heatmap.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
