% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdensity_map.R
\name{ggdensity_map}
\alias{ggdensity_map}
\title{Plots a density color map from a matrix or a molten data.frame.}
\usage{
ggdensity_map(
  m,
  rn.col = "rn",
  var.col = "variable",
  val.col = "value",
  grp.col = "group",
  col.map = "plasma",
  from = NULL,
  to = NULL,
  sort.fun = NULL,
  facet.by = "rows",
  ncores = 1,
  verbose = FALSE
)
}
\arguments{
\item{m}{Can be a \code{matrix} of numerical values or a molten
\code{data.frame} (e.g. that you can generate with
\code{data.table::melt()}) with a minimum of 4 columns:
\itemize{
 \item{1 column for rownames.}
 \item{1 column for the variable.}
 \item{1 column for values.}
 \item{and at least 1 column specifying the groups to which
 values belong to.}
}
The 'variable' and 'groups' columns can be as factors. If so,
the order of levels will influence the order of columns and
panels on the resulting plot.}

\item{rn.col}{A \code{character} when m is a molten data.frame to specify
the name of the rownames column.}

\item{var.col}{A \code{character} when m is a molten data.frame to specify
the name of the variable column.}

\item{val.col}{A \code{character} when m is a molten data.frame to specify
the name of the column containing values.}

\item{grp.col}{A \code{character} when m is a molten data.frame to specify
the name of the groups column.}

\item{col.map}{A \code{character} specifying a color map to use for the
density heatmap (Default: col.map = "plasma"; Supported:
col.map = c("A", "B", "C", "D", "E", "F", "G", "H")). For
more information about color maps see
\link[ggplot2]{scale_fill_viridis_c}.}

\item{from}{A \code{numeric} to specify the minimum value from which the
kernel estimation starts
(Default: from = min(m, na.rm = TRUE)).}

\item{to}{A \code{numeric} to specify the maximum value to which the
kernel estimation ends (Default: to = max(m, na.rm = TRUE)).}

\item{sort.fun}{A \code{character} matching a function to apply to the matrix
or the molten dataframe, to sort columns of the density
heatmap based on the values calculated
(Default: sort.fun = NULL ; Supported: "base::mean",
"base::median", "base::sd", "stats::IQR", ...and more).}

\item{facet.by}{A \code{character} specifying whether the density map on a
molten data.frame should be faceted on rows
(Default: facet.by = "rows") or on columns
(facet.by = "cols").}

\item{ncores}{An \code{integer} to specify the number of cores/threads to
be used to parallel-compute densities on matrix columns
(Default: ncores = 1).}

\item{verbose}{A \code{logical} to specify whether the function process
should be verbose (verbose = TRUE), or mute
(Default: verbose = FALSE).}
}
\value{
A \code{gg} plot of a density heatmap.
}
\description{
Plots a density color map from a matrix or a molten data.frame.
}
\examples{
# Create a density heatmap on the iris dataset:
ggdensity_map(m = as.matrix(iris[, 1:4])) +
  theme(axis.text.x = element_text(size = 12))
# Sort density heatmap columns using a function (here mean() from base):
ggdensity_map(m = as.matrix(iris[, 1:4]), sort.fun = "base::mean") +
  theme(axis.text.x = element_text(size = 12))
# Facetting the density heatmap on the iris dataset by Species:
dt.iris <- as.data.table(iris)
molten.iris <- melt.data.table(data = dt.iris, id.vars = "Species")
molten.iris[, rn := .I]
ggdensity_map(m = molten.iris, grp.col = "Species") +
  theme(axis.text.x = element_text(size = 12))
# Plot densities on a specific range of values:
ggdensity_map(m = molten.iris, grp.col = "Species", from = 2, to = 5) +
  theme(axis.text.x = element_text(size = 12))
# Change the color map:
ggdensity_map(
  m = molten.iris, grp.col = "Species", col.map = "D", from = 2, to = 5) +
  theme(axis.text.x = element_text(size = 12))
# Facetting the density heatmap by columns:
ggdensity_map(m = molten.iris, grp.col = "Species", facet.by = "cols") +
  theme(axis.text.x = element_text(size = 12, angle = -90))
}
\author{
Yoann Pageaud.
}
