% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva.R
\name{ggeigenvector}
\alias{ggeigenvector}
\title{Creates an eigenvector plot using ggplot2.}
\usage{
ggeigenvector(data, xcol, ycol, eigenvalues, colors, label = TRUE, title)
}
\arguments{
\item{data}{A \code{data.frame} containg:
N eigen vectors, one vector by column, with column names
numbered from "V1" to "VN",
a column "Groups" for the variable groups,
and a column "Labels" for the variable names.}

\item{xcol}{A \code{character} matching the column name of an
eigenvector, for which the values will be used as X axis
coordinates.}

\item{ycol}{A \code{character} matching the column name of an
eigenvector, for which the values will be used as Y axis
coordinates.}

\item{eigenvalues}{A \code{double} vector containing the eigenvalues of the
matrix.}

\item{colors}{A \code{character} vector of colors for the eigenvectors.
The length of this vector has to match the number of
different groups existing.}

\item{title}{A \code{character} that will be used as a title for the
plot.}
}
\value{
A \code{gg} plot of the variables following the 2 eigenvectors
        selected.
}
\description{
Creates an eigenvector plot using ggplot2.
}
\author{
Yoann Pageaud.
}
