% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghistogram.R
\name{gghist}
\alias{gghist}
\title{Plots an histogram using ggplot2 from a numeric or character vector.}
\usage{
gghist(
  x,
  xmin = min(x, na.rm = TRUE),
  xmax = max(x, na.rm = TRUE),
  binwidth = NULL,
  nbins = NULL,
  ngrads = 10,
  round.grad = 1,
  bin.col = "#0570b0",
  show.annot = "none",
  label.size = 4,
  facet = NULL,
  frow = NULL,
  fcol = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{A \code{numeric} or \code{character} vector to be used for
plotting the histogram.}

\item{xmin}{A \code{numeric} specifying the minimum value of the
histogram distribution. All values below xmin will be
aggregated on the first histogram bin
(Default: xmin = min(x, na.rm = TRUE)).}

\item{xmax}{A \code{numeric} specifying the maximum value of the
histogram distribution. All values above xmax will be
aggregated on the last histogram bin
(Default: xmax = max(x, na.rm = TRUE)).}

\item{binwidth}{An \code{integer} specifying the width for the histogram
bins. If the binwidth is not a divisor of the entire
distribution, the closest divisor is used. By default, the
bins width is guessed internally.}

\item{nbins}{An \code{integer} specifying the total number of bins to be
displayed on the plot. The resulting plot may not reflect
the number of bins specified, but should be close.
Moreover, if the range of the plot display exceed the range
of the histogram distribution, the number of visible bins
with a count > 0, will be lower, because the breaks of bins
are computed on the plot display range.}

\item{round.grad}{An \code{integer} specifying the number significant digits
to be considered when calculating graduations
(Default: round.grad = 1).}

\item{bin.col}{A \code{character} matching a R color code to be used to
fill the histogram bins (Default: bin.col = "#0570b0").}

\item{show.annot}{A \code{character} specifying what annotations should be
displayed on the histogram (Supoorted: show.annot = c(
"cutoff", "median", "all", "none");
Default: show.annot = "none").}

\item{label.size}{A \code{integer} specifying the size of annotations'
labels (Default: label.size = 4).}

\item{facet}{A \code{numeric} or \code{character} vector of the same
length as x, optionally with n levels to be used to
generate n histograms in separate panels.}

\item{frow}{An \code{integer} to specify the number of rows to display
panels for the facet_wrap() display.}

\item{fcol}{An \code{integer} to specify the number of columns to
display panels for the facet_wrap() display.}

\item{verbose}{A \code{logical} to verbose details about the histogram
computation. It may be useful to understand how the
resulting histogram was drawn.}

\item{ngrad}{An \code{integer} specifying the number of graduations to
display on the X-axis of the plot (Default: ngrad = 10).}
}
\value{
A \code{gg} plot of an histogram.
}
\description{
Plots an histogram using ggplot2 from a numeric or character vector.
}
\examples{
# Create a numerical distribution
distrib <- rnorm(100, 5, 1)
# Draw the default gghist() histogram
gghist(x = distrib)
# Increase display range
gghist(x = distrib, xmin = 0, xmax = 10)
# Decrease distribution range and aggregate extremas
gghist(x = distrib, xmin = 4, xmax = 7)
# Set binwidth equals 2
gghist(x = distrib, xmin = 4, xmax = 7, binwidth = 2)
# Set nbins to 10 (drawing 9) bins
gghist(x = distrib, xmin = 4, xmax = 7, nbins = 10)
# If binwidth and nbins are both specified, nbins is ignored
gghist(x = distrib, binwidth = 1, nbins = 10)
# When dislay range exceed distribution range, the number of visible bins is
# inferior to the value of nbins specified
gghist(x = distrib, nbins = 50, xmin = 0, xmax = 10)
# Turn verbose on.
gghist(x = distrib, nbins = 10, verbose = TRUE)
# Display first valley cut-off and median of the distribution.
gghist(x = distrib, show.annot = "all")
# Change annotations' labels size
gghist(x = distrib, show.annot = "all", label.size = 3)
# Set another filling color to the histogram
gghist(x = distrib, bin.col = "orange")
# Facet the data into 4 panels A, B, C & D
gghist(x = distrib, facet = rep(x = c("A", "B", "C", "D"), each = 25))
# Display all facets onto 1 row
gghist(
    x = distrib, facet = rep(x = c("A", "B", "C", "D"), each = 25), frow = 1)
# Create a character distribution
distrib_char <- rep(x = LETTERS, round(runif(26, 1, 99)))
# Draw a default gghist() histogram on character data
gghist(x = distrib_char)
# Set nbins to 10 bins; The closest divisor of the distribution is used
# instead
gghist(x = distrib_char, nbins = 10)
# Set binwidth to 3; The closest divisor of the distribution is used instead
gghist(x = distrib_char, binwidth = 3)
# Turn verbose on to understand what happened to binwidth internally.
gghist(x = distrib_char, binwidth = 3, verbose = TRUE)
# Facet the character data into 2 panels 1 & 2
gghist(
    x = distrib_char, binwidth = 3,
    facet = rep(seq(2), length(distrib_char)/2))
# Display all facets onto 1 column
gghist(
    x = distrib_char, binwidth = 3,
    facet = rep(seq(2), length(distrib_char)/2), fcol = 1)
}
\author{
Yoann Pageaud.
}
