% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbipca.R
\name{ggloadings}
\alias{ggloadings}
\title{Draws loadings from a PCA.}
\usage{
ggloadings(ggbiplot, loadings.data, loadings.col)
}
\arguments{
\item{loadings.data}{A \code{data.table} containing data necessary to plot
loadings.}

\item{loadings.col}{A \code{character} specifying a color to be used for
loadings.}
}
\value{
A \code{ggplot2} plot of PCA loadings.
}
\description{
Draws loadings from a PCA.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
