% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{ggpanel.corr}
\alias{ggpanel.corr}
\title{Plots results of correlation test between a single variable and multiple
others as jittered scatter plot divided into 4 different panels.}
\usage{
ggpanel.corr(data, p.cutoff = 0.01, corr.label.cutoff, jitter.height = 0.4)
}
\arguments{
\item{data}{A \code{data.table} with 3 to 5 columns:
\itemize{
 \item{column 1 - strings to be used as labels for
 individual dots.}
 \item{column 2 - correlation values.}
 \item{column 3 - correlation p-values.}
 \item{column 4 (optional) - groups to be used for
 coloring the dots.}
 \item{column 5 (optional) - values to be used to
 define dots sizes. It can be the sample size used
 for the calculation of the correlation test.}
}}

\item{p.cutoff}{A \code{numeric} between 0 and 1 to be used as a
maximum cut-off on p-values
(Default: p.cutoff = 0.01).}

\item{corr.label.cutoff}{A \code{numeric} vector of 1 or 2 values, between -1
and 1 to be used as the minimum cut-off on
positive and negative correlation values.
\itemize{
 \item{If 1 value is given, it will be define as the
 minimum cut-off on absolute correlation values
 (positive and negative ones).}
 \item{If 2 values are given, the smalest one will
 be used as a maximum cut-off on negative
 correlation values. the biggest one will be used as
 a minimum cut-off on positive correlation values.
 The smalest value must be comprised between -1 and
 0. The biggest value must be comprised between 0
 and 1.}
}}

\item{jitter.height}{A \code{numeric} to specify the amount of vertical
jitter (Default: jitter.height = 0.4).}
}
\value{
A \code{gg} plot object with 4 panels:
        \itemize{
         \item{1 panel with significant positive correlation values.}
         \item{1 panel with significant negative correlation values.}
         \item{1 panel with non-significant positive correlation values.}
         \item{1 panel with non-significant positive correlation values.}
        }
        Each panel displays results as jittered scatter plots.
}
\description{
Plots results of correlation test between a single variable and multiple
others as jittered scatter plot divided into 4 different panels.
}
\author{
Yoann Pageaud.
}
