% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{ggsidebar.basic}
\alias{ggsidebar.basic}
\title{Draws a ggplot2 of a basic sidebar.}
\usage{
ggsidebar.basic(
  data,
  palette,
  annot.sep = 0,
  annot.cut = 0,
  lgd.ncol = 1,
  facet = NULL
)
}
\arguments{
\item{data}{A \code{data.table} with the column names 'Samples','.id'
and 'Groups' in this order.}

\item{palette}{A \code{character} vector containing R colors like a
palette.}

\item{annot.sep}{A \code{double} specifying the width of the separation
spaces between annotations (Default: annot.sep = 0).}

\item{annot.cut}{A \code{double} specifying the width of cuts separating
annotation cells (Default: annot.cut = 0).}

\item{lgd.ncol}{An \code{integer} specifying the number of columns to be
used to display a legend (Default: lgd.ncol = 1).}
}
\value{
A \code{gg} object of the basic sidebar (a 'geom_tile()').
}
\description{
Draws a ggplot2 of a basic sidebar.
}
\examples{
# Prepare the dataset
dt_cars <- as.data.table(mtcars, keep.rownames = "Samples")
dt_cars <- dt_cars[, c("Samples", "am", "gear", "carb"), ]
dt_cars <- melt.data.table(
    data = dt_cars, id.vars = "Samples", variable.name = ".id",
    value.name = "Groups")
dt_cars[, Groups := paste(Groups, .id)]
# Create palette (length matches number of unique value in column 'Groups')
pal <- rainbow(n = 11)
# Draw the most basic annotation sidebar with 1 annotation only ('am')
ggsidebar.basic(data = dt_cars[.id == "am"], palette = pal)
# Draw an annotation sidebar with multiple annotations
ggsidebar.basic(data = dt_cars, palette = pal)
# Custom the annotation sidebar using ggplot2
ggsidebar.basic(data = dt_cars, palette = pal) +
    scale_y_discrete(expand = c(0,0)) +
    scale_x_discrete(expand = c(0,0)) +
    theme(
        axis.title.y = element_blank(),
        axis.text.x = element_text(angle = -45, hjust = 0, size = 10))
# Separate annotation with a blank space
ggsidebar.basic(data = dt_cars, palette = pal, annot.sep = 0.1)
# Separate samples within annotations by a blank space
ggsidebar.basic(data = dt_cars, palette = pal, annot.cut = 0.3)
# Display legend keys on 2 columns instead of 1
ggsidebar.basic(data = dt_cars, palette = pal, lgd.ncol = 2)
# Specify 1 of the annotation as the facet of the sidebar
ggsidebar.basic(data = dt_cars, palette = pal, facet = "am")
}
\author{
Yoann Pageaud.
}
