% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{ggsidebar.full}
\alias{ggsidebar.full}
\title{Creates a colored side annotation bars in ggplot2.}
\usage{
ggsidebar.full(
  sample.names,
  annot.grps,
  annot.pal,
  annot.pos = "top",
  annot.sep = 0,
  annot.cut = 0,
  merge.lgd = FALSE,
  right = FALSE,
  lgd.name = "Legends",
  lgd.ncol = 1,
  theme_legend = NULL,
  theme_annot = ggplot2::theme(axis.text.x = ggplot2::element_text(size = 12),
    axis.text.y = ggplot2::element_text(size = 12)),
  set.x.title = NULL,
  set.y.title = NULL,
  dendro.pos = "top",
  facet = NULL
)
}
\arguments{
\item{sample.names}{A \code{character} vector of the labels to be used for
annotation.}

\item{annot.grps}{A \code{list} of vectors of groups to which variables
belongs for the annotation sidebars. vectors lengths
have to match the number of variables.}

\item{annot.pal}{A \code{vector} or a list of vectors containing colors
as characters for the annotation sidebars. The length of
vectors has to match the number of levels of vectors
listed in 'annot.grps'.
\itemize{
 \item{If annot.pal is a list: its length must match the
       length of the list provided to 'annot.grps'.}
 \item{If annot.pal is a vector: make sure that the
       levels content of annotations listed in
       'annot.grps' is the same, and that no annotation
       contains less or more levels than another one in
       'annot.grps'.}
}}

\item{annot.pos}{A \code{character} specifying the position of the
annotation sidebar.\cr Possible values are: 'top',
'left' or 'both'.}

\item{annot.sep}{A \code{double} specifying the width of the separation
spaces between annotations (Default: annot.sep = 0).}

\item{annot.cut}{A \code{double} specifying the width of cuts separating
annotation cells (Default: annot.cut = 0).}

\item{merge.lgd}{A \code{logical} to specify whether annotation legends
should be merged (annot.lgd.merge = TRUE) or remain
separated (annot.lgd.merge = FALSE)
(Default: annot.lgd.merge = FALSE).}

\item{right}{A \code{logical} to specify that Y-axis should be
displayed on the right side of the plot
(Default: right = FALSE).}

\item{lgd.name}{A \code{character} to specify a title to the legend of
the plot, only if 'merge.lgd' = TRUE
(Default: lgd.name = "Legends").}

\item{lgd.ncol}{An \code{integer} specifying the number of columns to be
used to display a legend (Default: lgd.ncol = 1).}

\item{theme_legend}{A ggplot2 \code{theme} to specify any theme parameter you
wish to custom on legends (Default: theme_legend = NULL).
For more information about how to define a theme, see
\link[ggplot2]{theme}.}

\item{theme_annot}{A ggplot2 \code{theme} to specify any theme parameter you
wish to custom on the annotation bar
(Default: theme_annot = NULL). For more information about
how to define a theme, see \link[ggplot2]{theme}.}

\item{set.x.title}{A \code{character}to be used as the title for the X axis.}

\item{set.y.title}{A \code{character}to be used as the title for the Y axis.}

\item{dendro.pos}{A \code{character} specifying the position of the
dendrogram (Supported: dendro.pos = c('top', 'left')).}

\item{facet}{A \code{character} matching an annotation name in
'annot.grps' to be used to split the sidebar in separate
panels following the annotation.}
}
\value{
A \code{list} of length 2:
        \itemize{
         \item{'sidebar' contains the colored sidebar plot.}
         \item{'legends' lists legends of the matching sidebar as \code{gg}
               objects.}
        }
}
\description{
Creates a colored side annotation bars in ggplot2.
}
\examples{
# Draw a default sidebar
annotbar <- ggsidebar.full(
    sample.names = rownames(mtcars),
    annot.grps = list("Carb" = paste(mtcars$carb, "carb")),
    annot.pal = rainbow(n = 6))
# Draw the sidebar and customize it with ggplot2
annotbar$sidebar +
    theme(
        axis.text.x = element_text(angle = 45, hjust = 0),
        plot.margin = margin(0, 3, 0, 0, "cm"))
# Draw the sidebar's first legend
grid::grid.draw(annotbar$legends[[1]])
}
\author{
Yoann Pageaud.
}
