% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackbar.R
\name{ggstackbar}
\alias{ggstackbar}
\title{Draws stacked barplots from an annotation table}
\usage{
ggstackbar(
  dt,
  cols = NULL,
  y_breaks = 11,
  pal = BiocompR::biopalette(name = "rcartocolor_safe", mute = TRUE),
  lgd_nrow = 35,
  rm.null = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dt}{A \code{data.table} or a \code{data.frame} containing 1
annotation per column.}

\item{cols}{An \code{integer} or a \code{character} vector matching
columns in 'dt'. If 'cols' is a vector of integers, integers
must match columns position in the table. If 'cols' is a
vector of characters, characters must match columns names.
If 'cols' is NULL (the default), all columns are retrieved
from 'dt'.}

\item{y_breaks}{An \code{integer} specifying the number of graduations on the
Y axis (Default: y_breaks = 11).}

\item{pal}{A \code{character} vector matching existing color codes to be
used as a palette for the stacked barplot.}

\item{lgd_nrow}{An \code{integer} specifying the number of rows on which the
legend keys should be displayed (Default: lgd_nrow = 35).}

\item{rm.null}{A \code{logical} specifying whether the NULL elements
generated in the final list by ggstackbar() should be
automatically removed (rm.null = TRUE), or kept in the final
result (Default: rm.null = FALSE).}

\item{verbose}{A \code{logical} specifying whether the function should
verbose information while executing
(Default: verbose = TRUE).}
}
\value{
A \code{gg} objects list, each element being an individual stacked
        barplot.
}
\description{
Draws stacked barplots from an annotation table
}
\examples{
# Basic use of ggstackbar
res_stackbar <- ggstackbar(dt = ggplot2::mpg)
# Simply explore results: to see the stacked barplot of manufacturers, do:
res_stackbar$manufacturer
# Only generate stacked barplots for columns 'manufacturer', 'model' & 'year'
ls_res <- ggstackbar(
    dt = ggplot2::mpg, cols = c('manufacturer', 'model', 'year'))
ls_res$manufacturer
# Increase the number of graduation on the Y axis
ls_res <- ggstackbar(
    dt = ggplot2::mpg, cols = c('manufacturer', 'model', 'year'),
    y_breaks = 21)
ls_res$manufacturer
# Set custom color palette
ls_res <- ggstackbar(
    dt = ggplot2::mpg, cols = c('manufacturer', 'model', 'year'),
    y_breaks = 21, pal = biopalette(name = "ggsci_Nature", mute = TRUE))
ls_res$manufacturer
# Set number of lines on which to display legend's keys to 5 lines
ls_res <- ggstackbar(
    dt = ggplot2::mpg, cols = c('manufacturer', 'model', 'year'),
    y_breaks = 21, pal = biopalette(name = "ggsci_Nature", mute = TRUE),
    lgd_nrow = 5)
ls_res$manufacturer
# Arrange the output result in a single plot
ls_res <- ggstackbar(
    dt = ggplot2::mpg, cols = c('manufacturer', 'model', 'year'),
    y_breaks = 21, pal = biopalette(name = "ggsci_Nature", mute = TRUE),
    lgd_nrow = 38)
egg::ggarrange(plots = ls_res, nrow = 1)
}
\author{
Yoann Pageaud.
}
