% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsunset.R
\name{ggsunset}
\alias{ggsunset}
\title{Draws a sunset plot showing the completeness of a dataset.}
\usage{
ggsunset(
  mat,
  title = "Number of rows without missing data",
  col.pal = c("red", "gold", "blue4"),
  horizontal = FALSE,
  reverse = FALSE,
  n.grad = 15,
  display.cutoff = 0.03,
  display.num.smpl = 0.01
)
}
\arguments{
\item{mat}{A \code{matrix}.}

\item{title}{A \code{character} to specify the title of your plot.}

\item{col.pal}{A \code{character} vector of length 3 matching R
colors to used as a palette for the plot. col.pal can
use the palettes provided by the \link{biopalette}
function.\cr
(Default: col.pal = biopalette(
name = "BiocompR_sunset", mute = TRUE)).}

\item{horizontal}{A \code{logical} to specify whether the plot should
be drawn vertically or horizontally\cr
(Default: horizontal = FALSE).}

\item{reverse}{A \code{logical} to specify whether the order
displayed of the bins should reversed\cr
(Default: reverse = FALSE).}

\item{n.grad}{An \code{integer} specifying the number of graduation
to use on the Y-Axis\cr(Default: n.grad = 15).}

\item{display.cutoff}{A \code{double} specifying the minimum height of
a bin to label the number of samples on it. Below
this cutoff the number of samples is displayed in the
side margin\cr
(Default: display.cutoff = 0.03).}

\item{display.num.smpl}{A \code{double} specifying the minimum height of a
bin to display the number of samples in the side
margin\cr(Default: display.num.smpl = 0.01).}

\item{keep_2nd_ticks}{A \code{logical} to specify whether the ticks of the
secondary axis should be displayed or not\cr
(Default: keep_2nd_ticks = FALSE).}
}
\value{
A \code{gg} sunset plot.
}
\description{
Draws a sunset plot showing the completeness of a dataset.
}
\examples{
# Creating a matrix with randomly distributed 65\% NAs (MARs)
rand_vect <- rnorm(n = 100, mean = 0, sd = 1)
rand_vect[sample(100, 65)] <- NA
rand_mat <- matrix(data = rand_vect, ncol = 5)
# Basic sunset plot
ggsunset(mat = rand_mat)
# Change plot title on the left when the sunset is vertical
ggsunset(mat = rand_mat, title = "My new title")
# Change plot direction to horizontal (title now at the top)
ggsunset(mat = rand_mat, title = "My new title", horizontal = TRUE)
# Reverse the stacking
ggsunset(mat = rand_mat, reverse = TRUE)
# Change the number of graduations to something more meaningful
ggsunset(mat = rand_mat, n.grad = 5)
# Increase the label display cutoff to move labels on the right margin
ggsunset(mat = rand_mat, n.grad = 5, display.cutoff = 0.5)
# Increase the margin labels display cutoff to hide labels of smaller bins
# (This also automatically aggregate labels of contiguous smaller bins)
ggsunset(
    mat = rand_mat, n.grad = 5, display.cutoff = 0.5,
    display.num.smpl = 0.2)
# Change the color palette, e.g. using biopalette()
ggsunset(
    mat = rand_mat, n.grad = 5, reverse = TRUE,
    col.pal = biopalette(name = "Instagram", mute = TRUE))
# Free customization with ggplot2
ggsunset(
    mat = rand_mat, n.grad = 5, reverse = TRUE,
    col.pal = biopalette(name = "Instagram", mute = TRUE)) +
    theme(
        panel.border = element_rect( # Change plot border color
            color = "black", fill = "transparent", linewidth = 1),
        legend.position = "right", # plot legend on the right
        plot.margin = margin(0.3, 0, 0.3, 0.3, unit = "cm")) + # New margins
    guides( # Customize the colorbar legend
        fill = guide_colorbar(
            title = "N. columns", barheight = 5, barwidth = 1,
            ticks.linewidth = 0.5, ticks.colour = "black",
            frame.colour = "black"))
}
\author{
Yoann Pageaud.
}
