% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{ggvolcano.free}
\alias{ggvolcano.free}
\title{Plots any kind of results with P-values that can be displayed as a volcano
plot.}
\usage{
ggvolcano.free(
  data,
  label.cutoff = NULL,
  p.cutoff = 0.01,
  x.cutoff = NULL,
  title.x.cutoff = "X-Axis cutoff",
  x.col.sign = FALSE,
  force.label = NULL
)
}
\arguments{
\item{data}{A \code{data.table} with 3 to 5 columns:
\itemize{
 \item{column 1 - The labels.}
 \item{column 2 - X-axis values.}
 \item{column 3 - P-values.}
 \item{column 4 (optional) - groups to be used to color
 points. If none provided, you can use 'x.col.sign' to
 color dots based on the Y-axis category they belong
 to.}
 \item{column 5 (optional) - values to be used to
 define dots sizes based on another variable.}
}}

\item{label.cutoff}{A \code{numeric} vector of 1 or 2 values, to be used as
a cut-off on Y-axis values
(Default: label.cutoff = NULL does not show any label).
\itemize{
 \item{If 1 value is given, it will be defined as the
 minimum cut-off on absolute Y-axis values
 (positive and negative ones).}
 \item{If 2 values are given: The smallest one will be
 used as a maximum cut-off on negative Y-axis values.
 The biggest one will be used as a minimum cut-off on
 positive Y-axis values. The smallest value must be
 inferior or equal to 0. The biggest value must be
 superior or equal to 0.}
}
Dots outside these limits will be labeled. Dots within
the range of these limits will remain unlabeled.}

\item{p.cutoff}{A \code{numeric} between 0 and 1 to be used as a
maximum cut-off on p-values (Default: p.cutoff = 0.01).}

\item{x.cutoff}{A \code{numeric} vector of 1 or 2 values, to be used as
a cut-off on Y-axis values (Default: x.cutoff = NULL).
\itemize{
 \item{If 1 negative value is given, it will be defined
 as the maximum cut-off on negative Y-axis values.}
 \item{If 1 positive value is given, it will be defined
 as the minimum cut-off on positive Y-axis values.}
 \item{If 2 values are given: The smallest one will be
 used as a maximum cut-off on negative Y-axis values.
 The biggest one will be used as a minimum cut-off on
 positive Y-axis values. The smallest value must be
 inferior or equal to 0. The biggest value must be
 superior or equal to 0.}
}
x.cutoff value(s) will be used to draw vertical lines
on the volcano plot.}

\item{title.x.cutoff}{A \code{character} to specify how you wish to label the
Y-axis vertical lines on the volcano plot
(Default: title.cutoff = "X-Axis cutoff").}

\item{x.col.sign}{A \code{logical} to activate automatic coloring of data
based on their correlation category. If TRUE, elements
will be divided into 3 categories:
\itemize{
 \item{Blue - elements below the negative 'x.cutoff'.}
 \item{Red - elements above the positive 'x.cutoff'.}
 \item{Grey - any other element not meeting the
 requirements in the first nor the second categories.}
}
If FALSE, dots will be colored using groups from data
(if any provided). If no groups are passed within data,
dots will remain black.}

\item{force.label}{A \code{character} vector matching elements in the
first column of 'data' that you wish to label,
independently from their actual significance
(Default: force.label = NULL). force.label overrides
other labeling parameters.}
}
\value{
A \code{gg} volcano plot of your results.
}
\description{
Plots any kind of results with P-values that can be displayed as a volcano
plot.
}
\author{
Yoann Pageaud.
}
