% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr_ggvolcano.test.R
\name{ggvolcano.test}
\alias{ggvolcano.test}
\title{Plots results of statistical as volcano plot.}
\usage{
ggvolcano.test(
  data,
  p.cutoff = 0.01,
  fc.cutoff = 0,
  label = NULL,
  group.label = FALSE
)
}
\arguments{
\item{data}{A \code{data.table} with 3 to 5 columns:
\itemize{
 \item{column 1 - strings to be used as labels for
 individual dots.}
 \item{column 2 - log2 fold changes.}
 \item{column 3 - p-values.}
 \item{column 4 (optional) - groups to be used for
 coloring the dots. If not provided the dots will be
 split into two groups 1) those how meet the log2 fc.cutoff
 and p.cutoff, and 2) those who not.
 }
}}

\item{p.cutoff}{A \code{numeric} between 0 and 1 to be used as a
maximum cut-off on p-values
(Default: p.cutoff = 0.01).}

\item{fc.cutoff}{A \code{numeric} to be used as a cut-off on fold changes
(Default: fc.cutoff = 0.0).}

\item{label}{A \code{character} vector to be used to define which strings of
\code{data.table}, column 1 are displayed to label the dots.
  (Default: label = NULL).}

\item{group.label}{A \code{boolean} to be used to define whether labeling is done on
individual (label.level = FALSE) or group level (label.level = TRUE).
Parameter is only used if \code{data table} column 4 is set.
(Default: group.label = FALSE).}
}
\value{
A \code{gg} volcano plot object.
}
\description{
Plots results of statistical as volcano plot.
}
\author{
Verena Bitto.
}
