% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{heatmap_layout}
\alias{heatmap_layout}
\title{Builds the gg2heatmap layout using the egg package}
\usage{
heatmap_layout(
  dd.rows,
  dd.cols,
  show.annot,
  ddgr_seg_row,
  ddgr_seg_col,
  sidebar,
  htmp,
  legends,
  dend.row.size,
  dend.col.size,
  annot.size,
  lgd.layout,
  lgd.space.width
)
}
\value{
A \code{egg} object of the assembled heatmap.
}
\description{
Builds the gg2heatmap layout using the egg package
}
\author{
Yoann Pageaud.
}
\keyword{internal}
