% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{is.elt_blank}
\alias{is.elt_blank}
\title{Checks if a list's attributes has for class 'element_blank'.}
\usage{
is.elt_blank(arg)
}
\arguments{
\item{arg}{A \code{list}.}
}
\value{
A \code{logical}.
}
\description{
Checks if a list's attributes has for class 'element_blank'.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
