% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_plot.R
\name{ks.plot}
\alias{ks.plot}
\title{Computes pairwise Kolmogorov-Smirnov tests on a matrix and display results in
a fused plot}
\usage{
ks.plot(
  data,
  ncores = 1,
  order.select,
  order.method,
  axis.text.y,
  axis.text.x,
  annot.grps,
  annot.pal,
  annot.size,
  lgd.title,
  lgd.breaks1,
  lgd.breaks2,
  lgd.nbin1,
  lgd.nbin2,
  lgd.pal1,
  lgd.pal2
)
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Computes pairwise Kolmogorov-Smirnov tests on a matrix and display results in
a fused plot
}
\author{
Yoann Pageaud.
}
