% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{load.palettes}
\alias{load.palettes}
\title{Loads pre-defined palettes.}
\usage{
load.palettes()
}
\value{
Loads \code{character} vectors of colors to be used as palettes.
}
\description{
Loads pre-defined palettes.
}
\examples{
load.palettes()
}
\author{
Yoann Pageaud.
}
