% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stat_tests.R
\name{ls.psych.as.dt}
\alias{ls.psych.as.dt}
\title{Converts a list of type 'psych' objects into a data.table.}
\usage{
ls.psych.as.dt(psych.list)
}
\arguments{
\item{psych.list}{A \code{psych} list of correlation results.}
}
\value{
A \code{data.table} of the correlation results with columns:
        \itemize{
         \item{'var.name' - the name of the psych in the list.}
         \item{'cor' - the value of the correlation test.}
         \item{'pvalue' - the P-value associated to the correlation test.}
         \item{'nsample' - the number of samples used for the pairwise
         correlation.}
         \item{'stderr' - the standard error associated to the correlation
         test.}
        }
}
\description{
Converts a list of type 'psych' objects into a data.table.
}
\references{
\href{https://www.scholars.northwestern.edu/en/publications/psych-procedures-for-personality-and-psychological-research}{William R Revelle, psych: Procedures for Personality and Psychological Research. Northwestern University, Evanston, Illinois, USA (2017).}
}
\author{
Yoann Pageaud.
}
\keyword{internal}
