% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_polygons.R
\name{ls.quantile}
\alias{ls.quantile}
\title{Applies the quantile function on a list of vectors.}
\usage{
ls.quantile(ls, qtiles)
}
\arguments{
\item{ls}{A \code{list} of vectors.}

\item{qtiles}{A \code{double} vector specifying the values of percentiles of
interest. Values must be between 0 and 1.}
}
\value{
A \code{list} of vectors containing percentiles values, one vector by
distribution.
}
\description{
Applies the quantile function on a list of vectors.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
