% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_na.R
\name{manage.na}
\alias{manage.na}
\title{Keeps, removes or imputes missing values in a matrix or a data.frame based on
sample groups.}
\usage{
manage.na(data, method = "remove", groups)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} with column names.}

\item{method}{A \code{character} which defines the method to use for managing
NAs:
\itemize{
 \item{'keep' does nothing.}
 \item{'remove' removes all rows containing at least 1 NA.}
 \item{'impute' removes rows containing only NAs and impute
       missing values in remaining rows. The imputation uses
       the groups defined in the parameter 'groups': e.g. in
       row 1 if the value of sample_1 is NA, the missing value
       will be imputed using the median of values from other
       samples of the same group than sample_1.}
}}

\item{groups}{A vector of length ncol(data) specifying the groups to which
samples belong. Order of groups in the vector has to match the
order of column names in data to properly associate samples and
groups.}
}
\description{
Keeps, removes or imputes missing values in a matrix or a data.frame based on
sample groups.
}
\references{

}
\author{
Yoann Pageaud.
}
\keyword{internal}
