% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stat_tests.R
\name{metric_alias}
\alias{metric_alias}
\title{Returns the alias of a correlation test metric.}
\usage{
metric_alias(metric = "r")
}
\arguments{
\item{metric}{A \code{character} matching a specific metric return by the
\link[psych]{corr.test} function.}
}
\value{
A \code{character} being the alias of the metric given.
}
\description{
Returns the alias of a correlation test metric.
}
\examples{
# Get the alias of the metric 'r'
metric_alias(metric = "r")
}
\references{
\href{https://www.scholars.northwestern.edu/en/publications/psych-procedures-for-personality-and-psychological-research}{William R Revelle, psych: Procedures for Personality and Psychological Research. Northwestern University, Evanston, Illinois, USA (2017).}
}
\author{
Yoann Pageaud.
}
\keyword{internal}
