% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_plot.R
\name{min.max.update}
\alias{min.max.update}
\title{Checks if min & max are given by legend:
\itemize{
 \item{If given: updates min & max of the plot.}
 \item{If none: calculates min & max from melted triangle.}
}}
\usage{
\method{min}{max.update}(lgd.limits = NULL, melt.tri, tri.type)
}
\arguments{
\item{lgd.limits}{A \code{vector} of length 2, containing limit values.
Can be NULL\cr(Default: lgd.limits=NULL).}

\item{melt.tri}{A \code{data.frame} melted triangle containing test values.}

\item{tri.type}{A \code{character} specifying the type of triangle provided
\cr(Supported: tri.type = c("upper","lower")).}
}
\value{
A \code{list} of length 2 containing the updated min and max values.
}
\description{
Checks if min & max are given by legend:
\itemize{
 \item{If given: updates min & max of the plot.}
 \item{If none: calculates min & max from melted triangle.}
}
}
\author{
Yoann Pageaud.
}
\keyword{internal}
