% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stat_tests.R
\name{pairwise.ks}
\alias{pairwise.ks}
\title{Computes a Kolmogrov-Smirnov test between all columns of a data.frame.}
\usage{
pairwise.ks(data, statistic = "stat", ncores = 1)
}
\arguments{
\item{data}{A \code{data.frame} of numerical values.}

\item{statistic}{A \code{character} specifying the type of statistic to
retrieve from the test\cr
(Supported: statistic = c('n','stat','p')).}

\item{ncores}{An \code{integer} to specify the number of cores/threads to
be used to parallel-run tests.}
}
\value{
A \code{list} containing a matrix of the statistic values, a
        data.frame of the pairwise KS raw test results, and a table of
        pairwise combinations.
}
\description{
Computes a Kolmogrov-Smirnov test between all columns of a data.frame.
}
\examples{
# Format the airquality dataset
airquality <- data.table::as.data.table(airquality)
airquality[, Month:= factor(Month, labels = month.abb[5:9])]
df_month <- as.data.frame(data.table::dcast(
    data = airquality, Day~Month, value.var = "Ozone")[, c(2:6)])
# Do a pairwise Kolmogorov-Smirnov test and collect the KS 'stat' distances
# (only 1 core/thread used in the example)
ks_res <- pairwise.ks(data = df_month)
# Access the matrix of KS distances
ks_res$res.statistic
}
\author{
Yoann Pageaud.
}
