% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plots.R
\name{plot_asso_all_annot}
\alias{plot_asso_all_annot}
\title{Draws association test results between all columns from a data.frame.}
\usage{
plot_asso_all_annot(
  annot.table,
  perm.count = 10000,
  cohort.name = "dataset",
  verbose = FALSE
)
}
\arguments{
\item{annot.table}{A \code{data.frame} containing annotations, 1 annotation
per column.}

\item{perm.count}{An \code{integer} specifying the number of permutations
to realize on a vector, for the permutations matrix
initialization, to be used for calculating the
significance of a correlation test
(Default: perm.count = 10000).}

\item{cohort.name}{A \code{character} string specifying the name you want to
give to your dataset, that will be included in the plot
title (Default: cohort.name = "dataset").}

\item{verbose}{A \code{logical} to display information about the
step-by-step processing of the data if TRUE
(Default: verbose = FALSE).}
}
\value{
A \code{gg} plot.
}
\description{
Draws association test results between all columns from a data.frame.
}
\examples{
# Create mtcars annotation table
mtcars_annot <- data.table::as.data.table(
    mtcars[, c("cyl", "vs", "am", "gear", "carb")], keep.rownames = "cars")
# Plot association tests' results between all annotations
plot_asso_all_annot(
    annot.table = mtcars_annot, cohort.name = "mtcars_annot", verbose = TRUE)
}
\author{
Yoann Pageaud.
}
