% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_plots.R
\name{plot_asso_annot_PC}
\alias{plot_asso_annot_PC}
\title{Plots association tests' results between some annotations and some PCs.}
\usage{
plot_asso_annot_PC(
  annot.table,
  prcomp.res,
  perm.count = 10000,
  max.PCs = 8,
  dataset.name = "dataset",
  PC.origin = NULL,
  verbose = FALSE
)
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Plots association tests' results between some annotations and some PCs.
}
\examples{
# Create mtcars annotation table
mtcars_annot <- data.table::as.data.table(
    mtcars[, c("cyl", "vs", "am", "gear", "carb")], keep.rownames = "cars")
# Create mtcars matrix with scaled and transposed data
mat_mtcars <- scale(as.matrix(
    mtcars[, c("mpg", "disp", "hp", "drat", "wt", "qsec")]))
# Compute PCA on mtcars matrix
pcs_mtcars <- prcomp(x = mat_mtcars)
# Plot association tests' results between annotations and PCs
plot_asso_annot_PC(
    annot.table = mtcars_annot, prcomp.res = pcs_mtcars,
    dataset.name = "mtcars", PC.origin = "mpg, disp, hp, drat, wt & qsec",
    verbose = TRUE)
}
\author{
Yoann Pageaud.
}
