% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopalette.R
\name{plot_col}
\alias{plot_col}
\title{Plots a single color palette as a color bar.}
\usage{
plot_col(dt)
}
\arguments{
\item{dt}{A \code{data.table} subsetted from the \link{biopalette} function
internal palettes data.table.}
}
\value{
A \code{gg} plot object showing the colors of the palette
        (a 'geom_tile()').
}
\description{
Plots a single color palette as a color bar.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
