% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopalette.R
\name{plot_palette}
\alias{plot_palette}
\title{Plots all palettes from a given palette data.table object.}
\usage{
plot_palette(dt)
}
\arguments{
\item{dt}{A \code{data.table} subsetted from the \link{biopalette} function
internal palettes data.table.}
}
\value{
An \code{egg} plot of all color palettes contained in the input
        data.table.
}
\description{
Plots all palettes from a given palette data.table object.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
