% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{prepare_plot_export}
\alias{prepare_plot_export}
\title{Prepares gg2heatmap() results}
\usage{
prepare_plot_export(
  return_plots,
  dd.rows,
  dd.cols,
  show.annot,
  ddgr_seg_row,
  ddgr_seg_col,
  col_sidebar,
  htmp,
  htmp_legend,
  final.plot
)
}
\value{
A \code{list} of grobs and gg objects composing the heatmap.
}
\description{
Prepares gg2heatmap() results
}
\author{
Yoann Pageaud.
}
\keyword{internal}
