% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{raster.gg2grob}
\alias{raster.gg2grob}
\title{Rasterize a gg plot into a raster grob.}
\usage{
raster.gg2grob(gg.plot, filter = "Lanczos", size = 1080)
}
\arguments{
\item{gg.plot}{A \code{gg} plot to be rasterized.}

\item{filter}{A \code{character} to be used as a filter for ggplot
rasterization. The list of the supported rasterization
filters is available in magick::filter_types()
(Default: raster = "Lanczos"). Warning: Be aware that
rasterization may take several minutes to process the ggplot.}

\item{size}{An \code{integer} specifying the size of a squared raster in
pixels. If size = 1080 -> raster = 1080x1080
(Default: size = 1080).}
}
\value{
A \code{grob} of the rasterized ggplot.
}
\description{
Rasterize a gg plot into a raster grob.
}
\author{
Yoann Pageaud.
}
