% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{raster.ggplot.to.grob}
\alias{raster.ggplot.to.grob}
\title{Rasterize a gg plot into a raster grob.}
\usage{
raster.ggplot.to.grob(gg.plot, filter = "Lanczos")
}
\arguments{
\item{gg.plot}{A \code{gg} plot to be rasterized.}

\item{filter}{A \code{character} to be used as a filter for ggplot
rasterization. The list of the supported rasterization
filters is available in magick::filter_types()
(Default: raster = "Lanczos"). Warning: Be aware that
rasterization may take several minutes to process the ggplot.}
}
\value{
A \code{grob} of the rasterized ggplot.
}
\description{
Rasterize a gg plot into a raster grob.
}
\author{
Yoann Pageaud.
}
