% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbipca.R
\name{rename_dt_col}
\alias{rename_dt_col}
\title{Change colnames in annotation data.table to color, fill, or shape data.}
\usage{
rename_dt_col(data, color.data, fill.data, shape.data)
}
\arguments{
\item{data}{A \code{data.table} containing annotations matching data in
'prcomp.res'.}

\item{color.data}{A \code{character} specifying the column name in 'data' to
be used to map colors to points.}

\item{fill.data}{A \code{character} specifying the column name in 'data' to
be used to map filling colors to points (Important: this
parameter will only work if associated with 'shape.data',
and only if input values provided to scale_shape_manual()
are comprised between 21 and 25).}

\item{shape.data}{A \code{character} specifying the column name in 'data' to
be used to map shapes to points.}
}
\value{
A \code{data.table} with all annotations, and the colnames necessary
        for color, fill, or shape, changed.
}
\description{
Change colnames in annotation data.table to color, fill, or shape data.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
