% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{resize.grobs.oneway}
\alias{resize.grobs.oneway}
\title{Resizes heights or widths of a grob based on the dimensions of another grob}
\usage{
resize.grobs.oneway(grob1, grob2, dimensions)
}
\arguments{
\item{grob1}{A \code{grob} to be modified.}

\item{grob2}{A \code{grob} to be used as reference for dimensions
modifications.}

\item{dimensions}{A \code{character} specifying the type of dimensions to
resize, either 'heights' or 'widths'.}
}
\value{
A \code{grob} for which the dimensions have been modified.
}
\description{
Resizes heights or widths of a grob based on the dimensions of another grob
}
\author{
Yoann Pageaud.
}
