% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{resize.grobs}
\alias{resize.grobs}
\title{Resizes heights or widths of multiple grobs based on a given grob dimensions.}
\usage{
resize.grobs(ls.grobs, dimensions, start.unit, end.unit)
}
\arguments{
\item{ls.grobs}{A \code{grob} list. The list can be named if necessary.}

\item{dimensions}{A \code{character} specifying the type of dimensions to
resize, either 'heights' or 'widths'.}

\item{start.unit}{An \code{integer} specifying at which rank of the unit
object the dimension comparison between grobs should start.}

\item{end.unit}{An \code{integer} specifying at which rank of the unit
object the dimension comparison between grobs should end.}
}
\value{
A \code{grob} list, all resized with their dimensions modified by the
        unit.pmax() function.
}
\description{
Resizes heights or widths of multiple grobs based on a given grob dimensions.
}
\references{
\href{https://github.com/tidyverse/ggplot2/wiki/Align-two-plots-on-a-page}{Align two plots on a page - Mara Averick}
}
\author{
Yoann Pageaud.
}
