% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_na.R
\name{row.impute.na}
\alias{row.impute.na}
\title{Imputes missing data in a matrix row.}
\usage{
row.impute.na(data, miss.mat, grp_tbl, nr, r, impute.fun, vinterpolate)
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Imputes missing data in a matrix row.
}
\references{

}
\author{
Yoann Pageaud.
}
\keyword{internal}
