% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stat_tests.R
\name{simplecap}
\alias{simplecap}
\title{Converts to uppercase the first letter in a character string.}
\usage{
simplecap(x)
}
\arguments{
\item{x}{A \code{character}.}
}
\value{
A \code{character} with the first letter converted to uppercase.
}
\description{
Converts to uppercase the first letter in a character string.
}
\examples{
simplecap(x = "test")
}
\references{
\href{https://www.R-project.org/}{R Core Team (2022). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria.}
}
\author{
Yoann Pageaud.
}
\keyword{internal}
