% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{stack.grobs.legends}
\alias{stack.grobs.legends}
\title{Stack grobs legends vertically in separate spaces of specific heights}
\usage{
stack.grobs.legends(grobs.list, annot.grps, height.lgds.space)
}
\arguments{
\item{grobs.list}{A \code{list} of legends as grid objects.}

\item{annot.grps}{A \code{factor} list mapped to the legends.}

\item{height.lgds.space}{An \code{integer} specifying the height of the total
space which is supposed to gather all legends in
grobs.list.}
}
\value{
A \code{grob} containing all legends with their respective heights.
}
\description{
Stack grobs legends vertically in separate spaces of specific heights
}
\author{
Yoann Pageaud.
}
\keyword{internal}
