% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{sunset}
\alias{sunset}
\title{Draws a sunset plot showing the completeness of a dataset.}
\usage{
sunset(
  mat,
  title = "Number of rows without missing data",
  col.pal = c("red", "gold", "blue4"),
  horizontal = FALSE,
  reverse = FALSE,
  keep_2nd_ticks = FALSE,
  n.grad = 15,
  display.cutoff = 0.03,
  display.num.smpl = 0.01,
  lgd.pos = "bottom"
)
}
\arguments{
\item{mat}{A \code{matrix}.}

\item{title}{A \code{character} to specify the title of your plot.}

\item{col.pal}{A \code{character} vector of length 3 matching R
colors to used as a palette for the plot.
Alternatively, col.pal can use the palettes provides
within the script of the function:
pal_sunset, pal_westworld, pal_startrek,
pal_margesimpson, pal_eighties, pal_insta\cr
(Default: col.pal = pal_sunset).}

\item{horizontal}{A \code{logical} to specify whether the plot should
be drawn vertically or horizontally\cr
(Default: horizontal = FALSE).}

\item{reverse}{A \code{logical} to specify whether the order
displayed of the bins should reversed\cr
(Default: reverse = FALSE).}

\item{keep_2nd_ticks}{A \code{logical} to specify whether the ticks of the
secondary axis should be displayed or not\cr
(Default: keep_2nd_ticks = FALSE).}

\item{n.grad}{An \code{integer} specifying the number of graduation
to use on the Y-Axis\cr(Default: n.grad = 15).}

\item{display.cutoff}{A \code{double} specifying the minimum height of
a bin to label the number of samples on it. Below
this cutoff the number of samples is displayed in the
side margin\cr
(Default: display.cutoff = 0.03).}

\item{display.num.smpl}{A \code{double} specifying the minimum height of a
bin to display the number of samples in the side
margin\cr(Default: display.num.smpl = 0.01).}
}
\value{
A \code{gg} sunset plot.
}
\description{
Draws a sunset plot showing the completeness of a dataset.
}
\author{
Yoann Pageaud.
}
