% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stat_tests.R
\name{test.annots}
\alias{test.annots}
\title{Tests association of an annotation with another one or with a PC.}
\usage{
test.annots(x, y, perm.matrix = NULL)
}
\arguments{
\item{x}{A \code{vector} of numerics or factors containing data
from one annotation.}

\item{y}{A \code{vector} of numerics or factors containing data
from a PCA principal component or another annotation.}

\item{perm.matrix}{A \code{matrix} containing permutations of the order of
rows, as integers, based on annotations length. The
permutation matrix can optionally be used to test the
significance of a correlation value between 2 annotations
or an annotation and a PCA principal component.}
}
\value{
A \code{data.table} containing all results from an association test
        between x and y.
}
\description{
Tests association of an annotation with another one or with a PC.
}
\examples{
# Most basic statistic test between 2 vectors of integers
test.annots(x = 1:15, y = 24:10)
}
\author{
Yoann Pageaud.
}
