% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stat_tests.R
\name{test_asso_all_annot}
\alias{test_asso_all_annot}
\title{Tests associations between all annotations stored in a data.frame.}
\usage{
test_asso_all_annot(annot.table, perm.count = 10000, verbose = FALSE)
}
\arguments{
\item{annot.table}{A \code{data.frame} containing all annotations, 1
annotation per column.}

\item{perm.count}{An \code{integer} specifying the number of permutations to
realize on a vector, for the permutations matrix
initialization, to be used for calculating the significance
of a correlation test (Default: perm.count = 10000).}

\item{verbose}{A \code{logical} to display information about the
step-by-step processing of the data if TRUE
(Default: verbose = FALSE).}
}
\value{
A \code{data.table} containing all association test results.
}
\description{
Tests associations between all annotations stored in a data.frame.
}
\examples{
# Create mtcars annotation table
mtcars_annot <- data.table::as.data.table(
    mtcars[, c("cyl", "vs", "am", "gear", "carb")], keep.rownames = "cars")
# Test association between all annotations
asso_res <- test_asso_all_annot(
    annot.table = mtcars_annot, perm.count = 10000, verbose = TRUE)
# Table summarizing tests' results
asso_res
}
\author{
Yoann Pageaud.
}
