% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stat_tests.R
\name{test_asso_annot_pc}
\alias{test_asso_annot_pc}
\title{Tests associations between a set of annotations and PCs from a prcomp object.}
\usage{
test_asso_annot_pc(
  annot.table,
  prcomp.res,
  perm.count = 10000,
  max.PCs = 8,
  verbose = FALSE
)
}
\arguments{
\item{annot.table}{A \code{data.frame} containing all annotations, 1
annotation per column.}

\item{prcomp.res}{A PCA result of classes \code{prcomp} or
\code{irlba_prcomp} resulting from stats::prcomp() or
irlba::prcomp_irlba().}

\item{perm.count}{An \code{integer} specifying the number of permutations to
realize on a vector, for the permutations matrix
initialization, to be used for calculating the significance
of a correlation test (Default: perm.count = 10000).}

\item{max.PCs}{An \code{integer} specifying the maximum number of
principal components to consider for association tests with
annotations (Default: max.PCs = 8).}

\item{verbose}{A \code{logical} to display information about the
step-by-step processing of the data if TRUE
(Default: verbose = FALSE).}
}
\value{
A \code{data.table} containing all association test results.
}
\description{
Tests associations between a set of annotations and PCs from a prcomp object.
}
\examples{
# Create mtcars annotation table
mtcars_annot <- data.table::as.data.table(
    mtcars[, c("cyl", "vs", "am", "gear", "carb")], keep.rownames = "cars")
# Create mtcars matrix with scaled and transposed data
mat_mtcars <- scale(as.matrix(
    mtcars[, c("mpg", "disp", "hp", "drat", "wt", "qsec")]))
# Compute PCA on mtcars matrix
pcs_mtcars <- prcomp(x = mat_mtcars)
# Test association between annotations and principal components
asso_res <- test_asso_annot_pc(
    annot.table = mtcars_annot, prcomp.res = pcs_mtcars, verbose = TRUE)
# Table summarizing tests' results
asso_res
}
\author{
Yoann Pageaud.
}
