% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbipca.R
\name{theme_biplot}
\alias{theme_biplot}
\title{ggplot2 theme for PCA biplot functions.}
\usage{
theme_biplot()
}
\value{
A \code{ggplot2} object themed for biplot functions.
}
\description{
ggplot2 theme for PCA biplot functions.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
