% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{update_guide_colorbar}
\alias{update_guide_colorbar}
\title{Update a guide_colorbar with values from a new one, with exceptions.}
\usage{
update_guide_colorbar(
  old_guide = guide_colorbar(),
  new_guide,
  forced_param = NULL
)
}
\arguments{
\item{old_guide}{A \code{guide} of class 'colorbar' you want to update.}

\item{new_guide}{A \code{guide} of class 'colorbar' updating the old_guide
with its values.}

\item{forced_param}{A named \code{list} containing guide_colorbar parameters
associated with their values to be conserved.
Independently from whether the there are differences or
not between old and new guides, values passed using
'forced_param' will override old and new guide values.}
}
\value{
An updated \code{guide} of class 'colorbar'.
}
\description{
Update a guide_colorbar with values from a new one, with exceptions.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
