% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeva.R
\name{var.accounted}
\alias{var.accounted}
\title{Calculate the variance accounted of an eigenvalue.}
\usage{
var.accounted(colname, eigen.values)
}
\arguments{
\item{colname}{A \code{character} matching a column name.}

\item{eigen.values}{A \code{numeric} vector of all eigenvalues.}
}
\value{
A \code{character} containg the value of the eigenvalue variance
        accounted for its associated eigen vector.
}
\description{
Calculate the variance accounted of an eigenvalue.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
