% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings_handling.R
\name{warn.handle}
\alias{warn.handle}
\title{Filters irrelevant warnings matching a regular expression.}
\usage{
warn.handle(pattern, ...)
}
\arguments{
\item{pattern}{A \code{character} to be processed as a regular expression.}
}
\value{
A \code{character} list of warning messages without the warnings
        matching the pattern.
}
\description{
Filters irrelevant warnings matching a regular expression.
}
\author{
Yoann Pageaud.
}
