% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{basic.ggplot.tri}
\alias{basic.ggplot.tri}
\title{Draws a ggplot for a basic upper or lower triangle.}
\usage{
basic.ggplot.tri(
  melt.tri,
  grid.col,
  grid.thickness,
  lgd.title,
  lgd.text,
  lgd.pal,
  min_tri,
  max_tri,
  lgd.breaks,
  lgd.round,
  lgd.ticks,
  lgd.nbin,
  lgd.height,
  lgd.width,
  rasteri,
  lgd.ticks.linewidth,
  lgd.frame.col,
  lgd.frame.linewidth,
  diag.col,
  set.lgd.title
)
}
\arguments{
\item{melt.tri}{A \code{data.frame} melted triangle containing test
values.}

\item{grid.col}{A \code{character} specifying the color of the grid.}

\item{grid.thickness}{A \code{double} value for the thickness of the grid.}
}
\value{
A \code{gg} object of a basic triangle plot (a 'geom_tile()').
}
\description{
Draws a ggplot for a basic upper or lower triangle.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
