% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopalette.R
\name{biopalette}
\alias{biopalette}
\title{A color palette advisor for biology plots.}
\usage{
biopalette(x = "all", type = "all", length = 1)
}
\arguments{
\item{x}{A \code{character} vector of keywords. Currently supported
keywords are: "all", "color blind", "protanopia",
"deuteranopia", "viridis", "biocompr", "tritanopia",
"methylation", "expression", "test vs control",
"chromatin accessibility", "completeness" and "bonus". You can
pass multiple keywords to x as a character vector:
e.g. x = c("color blind", "deuteranopia", "methylation") will
return the "BiocompR_deflt", "viridis_D_viridis" and
"viridis_E_cividis" palettes.}

\item{type}{A \code{character} matching the data type you are working with.
Supported types are: c("sequential", "diverging",
"qualitative").}

\item{length}{An \code{integer} specifying the minimum number of colors
palettes should contain.}
}
\value{
A \code{data.table} of the queried palettes.
}
\description{
A color palette advisor for biology plots.
}
\examples{
#To show all palettes available
biopalette()
}
\author{
Yoann Pageaud.
}
