% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopalette.R
\name{biopalette}
\alias{biopalette}
\title{A color palette advisor for biology plots.}
\usage{
biopalette(x = NULL, name = NULL, type = NULL, show.pal = FALSE, mute = FALSE)
}
\arguments{
\item{x}{A \code{character} vector of keywords.\cr
Supported keywords can be of 4 types:
\itemize{
 \item{A package name: e.g. 'biocompr', 'ggsci', 'viridis',
 ...}
 \item{A biology or clinic data type: e.g. 'methylation',
 'test vs control', 'expression', 'genomic annotations',
 'tumor type', ...}
 \item{A type of color blindness: e.g. 'protanopia',
 'deuteranopia', 'tritanopia', 'color blind'.}
 \item{A consortium name: 'PCAWG'.}
}
All available keywords are accessible in the biopalette
internal data.table, by using biopalette() without any
argument.
You can pass multiple keywords to x as a character vector:
e.g. x = c("protanopia","methylation","biocompr") will return
2 BiocompR methylation palettes protanopia safe.}

\item{name}{A \code{character} matching a specific palette name.}

\item{type}{A \code{character} matching the data type you are working
with. Supported types are:
type = c("sequential", "diverging", "qualitative").}

\item{show.pal}{A \code{logical} to plot color palettes returned by your
query (show.pal = TRUE) or not (Default: show.pal = FALSE).}

\item{mute}{A \code{logical} to turn off (Default: mute = FALSE) or on
(mute = TRUE) palette name display when looking for a palette
using keywords. A palette name is not displayed when you call
it using the 'name' argument.}
}
\value{
When x = NULL and name = NULL, the whole palette \code{data.table}
        stored in biopalette() is returned.\cr
        When 1 matching palette is found, it is returned as a
        \code{character} vector.\cr
        When multiple matches are found, matches are returned as a
        \code{list} of palettes.
}
\description{
A color palette advisor for biology plots.
}
\examples{
# To print the internal data.table of all palettes available in biopalette()
# and to plot all palettes.
biopalette(show.pal = TRUE)
# To list all palettes from BiocompR
biopalette(x = "biocompr")
# To plot palettes matching a given query
biopalette(x = "biocompr", show.pal = TRUE)
# To list all palettes suiting for methylation data
biopalette(x = "methylation")
# To list color blind safe palettes
biopalette(x = "color blind")
# Not all palettes are fully colorblind safe, so you can select palette safe
# for a specific color blindness (protanopia, deuteranopia or tritanopia).
biopalette(x = "deuteranopia")
# To list color blind safe palettes from BiocompR suiting for methylation
# data
biopalette(x = c("color blind", "biocompr", "methylation"))
# biopalette() also supports partial match with supported keywords
# e.g. 'expr' for 'expression'
biopalette(x = "expr")
# If keywords return a single palette you can mute the palette name printed
biopalette(x = c("protanopia","tumor type","ggsci"), mute = TRUE)
# To get a palette using its name: e.g. the 'Okabe-Ito' palette
biopalette(name = "Okabe-Ito")
# To list palettes suiting specific data type: e.g. qualitative data
biopalette(type = "qualitative")
# Use case of biopalette() in a ggplot2 figure
ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) +
    geom_point() +
    scale_color_manual(values = biopalette(name = "BiocompR_sunset"))
}
\references{
\href{https://colorbrewer2.org/}{RColorBrewer - Erich Neuwirth}

\href{https://nanx.me/ggsci/}{Xiao N (2023). ggsci: Scientific Journal and Sci-Fi Themed Color Palettes for 'ggplot2'.}

\href{https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html}{Garnier, Simon, Ross, Noam, Rudis, Robert, Camargo, Pedro A, Sciaini, Marco, Scherer, Cédric (2023). viridis(Lite) - Colorblind-Friendly Color Maps for R. doi:10.5281/zenodo.4679423, viridis package version 0.6.4}

\href{https://cran.r-project.org/web/packages/dichromat/index.html}{Dichromat: Color Schemes for Dichromats - Thomas Lumley et al.}

\href{https://www.nature.com/articles/s41586-020-1969-6}{Pan-cancer analysis of whole genomes - The ICGC/TCGA Pan-Cancer Analysis of Whole Genomes Consortium}

\href{https://github.com/YoannPa/methview.qc}{Pageaud Y. et al., Visualize quality control data from methylation array dataset with methview.qc}

\href{https://jfly.uni-koeln.de/color/}{Color Universal Design (CUD) - How to make figures and presentations that are friendly to Colorblind people - Masataka Okabe & Kei Ito}

\href{https://jakubnowosad.com/rcartocolor/}{Nowosad, J. (2018). 'CARTOColors' Palettes. R package version 1.0.0.}

\href{https://academic.oup.com/bioinformatics/article/32/18/2847/1743594}{Gu Z, Eils R, Schlesner M (2016). “Complex heatmaps reveal patterns and correlations in multidimensional genomic data.” Bioinformatics. doi:10.1093/bioinformatics/btw313.}
}
\author{
Yoann Pageaud.
}
