% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{build.layout}
\alias{build.layout}
\title{Builds legends layout.}
\usage{
build.layout(col_table, height.lgds.space)
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Builds legends layout.
}
\references{

}
\author{
Yoann Pageaud.
}
\keyword{internal}
