% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{chk.dt}
\alias{chk.dt}
\title{Checks the data.table provided to ggpanel.corr() and build.ggvolcano().}
\usage{
chk.dt(data, data.type)
}
\arguments{
\item{data}{A \code{data.table} with 3 to 5 columns:
\itemize{
 \item{column 1 - The labels.}
 \item{column 2 - Y-axis values.}
 \item{column 3 - P-values.}
 \item{column 4 (optional) - groups to be used to color
 points. If none provided, you can use 'x.col.sign' to color
 dots based on the X-axis category they belong to.}
 \item{column 5 (optional) - values to be used to define
 dots sizes based on another variable.}
}}

\item{data.type}{A \code{character} to specify the type of test used to
generate 'data'.
(Supported: data.type = c("t.test", "corr")).}
}
\value{
A valid \code{data.table} with new column names.
}
\description{
Checks the data.table provided to ggpanel.corr() and build.ggvolcano().
}
\author{
Yoann Pageaud.
}
\keyword{internal}
