% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{chk.dt}
\alias{chk.dt}
\title{Checks the data.table provided to ggpanel.corr() and ggvolcano.corr().}
\usage{
chk.dt(data)
}
\arguments{
\item{data}{A \code{data.table} which contains correlation data.}
}
\value{
A \code{data.table} with new column names.
}
\description{
Checks the data.table provided to ggpanel.corr() and ggvolcano.corr().
}
\author{
Yoann Pageaud.
}
\keyword{internal}
