% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_plot.R
\name{fused.view}
\alias{fused.view}
\title{Displays 2 triangle matrices together as a fused plot.}
\usage{
fused.view(
  sample.names,
  upper.mat,
  lower.mat,
  order.select,
  order.method,
  hclust.method,
  annot.grps = list(Groups = seq(length(sample.names))),
  annot.pal = grDevices::rainbow(n = length(sample.names)),
  annot.pos = "top",
  annot.size = 1,
  annot.text = NULL,
  lgd.merge = FALSE,
  lgd.ncol = NULL,
  lgd.space.height = 26,
  lgd.space.width = 1,
  annot.split = FALSE,
  dendro.pos = "none",
  dendro.size = 0,
  grid_col = "white",
  grid_linewidth = 0.3,
  upper_theme = NULL,
  upper_scale_fill_grad = ggplot2::scale_fill_gradientn(colors =
    BiocompR::biopalette(name = "viridis_C_plasma"), name = "Upper\\ntriangle values"),
  upper_guide_custom_bar = ggplot2::guide_colorbar(barheight = 23, barwidth = 0.7,
    ticks.linewidth = 0.5, ticks.colour = "black", frame.linewidth = 0.5, frame.colour =
    "black"),
  lower_theme = NULL,
  lower_scale_fill_grad = ggplot2::scale_fill_gradientn(colors =
    BiocompR::biopalette(name = "viridis_D_viridis"), name = "Lower\\ntriangle values"),
  lower_guide_custom_bar = ggplot2::guide_colorbar(barheight = 0.7, barwidth = 23,
    ticks.linewidth = 0.5, ticks.colour = "black", frame.linewidth = 0.5, frame.colour =
    "black"),
  diagonal_col = "white",
  plot_title = NULL,
  axis.title = ggplot2::element_blank(),
  axis.title.x = ggplot2::element_blank(),
  axis.title.y = ggplot2::element_blank(),
  axis.text = ggplot2::element_text(size = 12),
  axis.text.x = ggplot2::element_text(angle = 90, hjust = 0, vjust = 0.5),
  axis.text.y = ggplot2::element_blank(),
  axis.ticks = ggplot2::element_line(color = "black"),
  lgd.title2 = ggplot2::element_blank(),
  lgd.text2 = ggplot2::element_blank(),
  lgd.breaks = NULL,
  lgd.breaks1 = 5,
  lgd.breaks2 = 5,
  lgd.labels = NULL,
  lgd.labels1 = NULL,
  lgd.labels2 = NULL,
  lgd.round = NULL,
  lgd.round1 = 2,
  lgd.round2 = 2,
  lgd.limits = NULL,
  lgd.limits1 = NULL,
  lgd.limits2 = NULL,
  lgd.ticks2 = NULL,
  lgd.ticks.linewidth2 = 2,
  lgd.nbin2 = 5,
  lgd.height2 = 1,
  lgd.width2 = 30,
  lgd.frame.col = "grey",
  lgd.frame.linewidth = 1.5,
  lgd.frame.linewidth1 = NULL,
  lgd.frame.linewidth2 = NULL,
  raster2 = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Displays 2 triangle matrices together as a fused plot.
}
\author{
Yoann Pageaud.
}
