% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_plot.R
\name{fused.view}
\alias{fused.view}
\title{Displays 2 matrices of results as a fused plot.}
\usage{
fused.view(
  sample.names,
  upper.mat,
  lower.mat,
  order.select,
  order.method,
  hclust.method,
  correlation.order,
  annot.grps = list(Groups = seq(ncol(data))),
  annot.pal = rainbow(n = ncol(data)),
  annot.pos = "top",
  annot.size = 0,
  annot.lgd.merge = FALSE,
  annot.split = TRUE,
  dendro.pos = "none",
  dendro.size = 0,
  grid.col = "grey",
  grid.thickness = 0.5,
  axis.title = element_blank(),
  axis.title.x = element_blank(),
  axis.title.y = element_blank(),
  axis.text = element_text(size = 12),
  axis.text.x = element_text(angle = 90, hjust = 0, vjust = 0.5),
  axis.text.y = element_blank(),
  axis.ticks = element_line(color = "black"),
  set.x.title = NULL,
  set.y.title = NULL,
  set.lgd1.title = NULL,
  set.lgd2.title = NULL,
  diag.col = "white",
  lgd.pal1 = NULL,
  lgd.pal2 = NULL,
  lgd.title = element_blank(),
  lgd.title1 = element_blank(),
  lgd.title2 = element_blank(),
  lgd.text = element_text(size = 12),
  lgd.text1 = element_blank(),
  lgd.text2 = element_blank(),
  lgd.breaks = NULL,
  lgd.breaks1 = NULL,
  lgd.breaks2 = NULL,
  lgd.labels = NULL,
  lgd.labels1 = NULL,
  lgd.labels2 = NULL,
  lgd.round = NULL,
  lgd.round1 = 2,
  lgd.round2 = 2,
  lgd.limits = NULL,
  lgd.limits1 = NULL,
  lgd.limits2 = NULL,
  lgd.ticks = TRUE,
  lgd.ticks1 = NULL,
  lgd.ticks2 = NULL,
  lgd.ticks.linewidth = 2,
  lgd.ticks.linewidth1 = NULL,
  lgd.ticks.linewidth2 = NULL,
  lgd.nbin = NULL,
  lgd.nbin1 = NULL,
  lgd.nbin2 = NULL,
  lgd.height1 = 30,
  lgd.height2 = 1,
  lgd.width1 = 1,
  lgd.width2 = 30,
  lgd.frame.col = "grey",
  lgd.frame.linewidth = 1.5,
  lgd.frame.linewidth1 = NULL,
  lgd.frame.linewidth2 = NULL,
  raster = FALSE,
  raster1 = NULL,
  raster2 = NULL,
  add.ggplot.arg = NULL
)
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Displays 2 matrices of results as a fused plot.
}
\author{
Yoann Pageaud.
}
