% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg2heatmap.R
\name{gg2heatmap}
\alias{gg2heatmap}
\title{WARNING: Unstable! Creates a custom heatmap with dendrograms and annotations.}
\usage{
gg2heatmap(
  m,
  na.handle = "remove",
  dist.method = "manhattan",
  rank.fun = NULL,
  top.rows = NULL,
  dendrograms = TRUE,
  imputation.grps = NULL,
  ncores = 1,
  heatmap.pal = c("steelblue", "gray95", "darkorange"),
  annot.grps = list(Groups = seq(ncol(m))),
  annot.pal = rainbow(n = ncol(m)),
  annot.size = 1,
  annot.lgd.space = 0,
  annot.split = FALSE,
  lgd.lab = "Legends",
  lgd.pos.x = 0.5,
  lgd.pos.y = 0.5,
  lgd.merge = FALSE
)
}
\arguments{
\item{m}{A \code{matrix}.}

\item{na.handle}{A \code{character} to specify how missing values
should be handled (Default: na.handle = 'remove';
Supported: na.handle = c('keep','impute','remove')).}

\item{dist.method}{A \code{character} vector to specify the distance
methods to be used on the matrix rows and columns.
\itemize{
 \item{If the vector is of length 1: the given method
       will be applied on rows and columns of the
       matrix.}
 \item{If the vector is of length 2: the first method
       will be applied on matrix rows, and the second
       method will be applied on matrix columns.}
}
(Default: dist.method = 'manhattan';
Supported: dist.method = c('euclidean', 'maximum',
'manhattan', 'canberra', 'binary', 'minkowski',
'none')).}

\item{rank.fun}{A \code{character} to specify a function to apply on
matrix rows to order them on the final heatmap. If
rank.fun = NULL the order of rows in the matrix will
be kept (Default: rank.fun = NULL;
Supported: rank.fun = 'sd').}

\item{top.rows}{An \code{integer} to specify the number of top rows
you want to display on the heatmap
(Default: top.rows = NULL).}

\item{dendrograms}{A \code{logical} vector to specify whether dendrograms
should be plotted with the heatmap.
\itemize{
 \item{If dendrograms = TRUE: both dendrograms on row
       and columns of the matrix will be plotted.}
 \item{If dendrograms = FALSE: both dendrograms on
       rows and columns of the matrix will NOT be
       plotted.}
 \item{If the vector is of length 2: the first logical
       will apply for rows, and the second logical
       will apply for columns.}
 \item{If 'rank.fun' and 'top.rows' are not NULL, and
       dendrograms on rows set to TRUE, the ranking of
       rows and the selection of the top ones will be
       made before the computation of dendrograms.}
}}

\item{imputation.grps}{A \code{character} vector defining groups to which
columns of the matrix belong. These groups are use to
make group-wise imputation of missing values between
columns. The vector has to be of the same length than
the number of columns in the matrix
(Default: imputation.grps = NULL).}

\item{ncores}{An \code{integer} to specify the number of
cores/threads to be used to parallel-compute distances
for dendrograms.}

\item{heatmap.pal}{A \code{character} vector to be used as a color
palette for the heatmap.}

\item{annot.grps}{A \code{list} of vectors of groups to which variables
belongs for the annotation sidebars. 'Vectors' lengths
have to match the number of variables.}

\item{annot.pal}{A \code{vector} or a list of vectors containing colors
as characters for the annotation sidebars. The length
of vectors has to match the number of levels of
vectors listed in 'annot.grps'. If a list is provided,
its length must match the length of the list provided
to 'annot.grps'.}

\item{annot.size}{A \code{numeric} defining the width of the annotation
bars (Default: annot.size = 1).}

\item{annot.lgd.space}{A \code{numeric} defining the size of the space
separating the different annotation bar legends
(Default: annot.lgd.space = 0).}

\item{annot.split}{A \code{logical} to specify whether a space should be
added between each annotation (annot.split = TRUE) or
not (annot.split = FALSE)
(Default: annot.split = FALSE).}

\item{lgd.lab}{A \code{character} specifying the name of annotation
side bar legends, when legends are merged
(lgd.merge = TRUE).}

\item{lgd.pos.x}{A \code{numeric} vector or unit object specifying the
x-location of the legends (Default: lgd.pos.x = 0.5).}

\item{lgd.pos.y}{A \code{numeric} vector or unit object specifying the
y-location of the legends (Default: lgd.pos.y = 0.5).}

\item{lgd.merge}{A \code{logical} specifying whether the legends of
multiple annotation bars should be merged
(lgd.merge = TRUE) or remain separated
(lgd.merge = FALSE). lgd.merge is especially useful
when you want to map the same color palette to
multiple annotations sharing the same values
(Default: lgd.merge = FALSE).}
}
\value{
A \code{grob} of a heatmap plotted with ggplot2.
}
\description{
WARNING: Unstable! Creates a custom heatmap with dendrograms and annotations.
}
\author{
Yoann Pageaud.
}
