% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggasycorr.R
\name{ggasycorr}
\alias{ggasycorr}
\title{Draws an asymmetrical pairwise correlation plot.}
\usage{
ggasycorr(
  psych_obj,
  pval_cutoff = 0.05,
  show_pval_below = 1,
  show_abs_cor_above = 0
)
}
\arguments{
\item{psych_obj}{A \code{psych} object containing correlation
results, generated by psych::corr.test().}

\item{pval_cutoff}{A \code{numeric} comprised within [0;1] specifying
the maximum cut-off for a P-value to be considered
significant (Default: pval_cutoff = 0.05).}

\item{show_pval_below}{A \code{numeric} comprised within [0;1] specifying
the maximum cut-off on P-values for results to be
displayed (Default: show_pval_below = 1).}

\item{show_abs_cor_above}{A \code{numeric} comprised within [0;1] specifying
the minimum cut-off on absolute correlation for
results to be displayed
(Default: show_abs_cor_above = 0).}
}
\value{
A \code{gg} asymmetrical pairwise correlation plot.
}
\description{
Draws an asymmetrical pairwise correlation plot.
}
\examples{
# Most basic asymetrical pairwise correlation plot using New York air quality
# measurements (Pearson's correlation with Holm's correction applied)
cor_res <- psych::corr.test(airquality[1:2], airquality[3:6])
ggasycorr(psych_obj = cor_res)
# There is a significant positive correlation between solar radiation
# intensities and temperatures measured.
# There is a significant negative correlation between ozone levels and
# wind speeds.
# There is a significant positive correlation between ozone levels and
# temperatures measured.
#
# Do Spearman's correlation instead
cor_res <- psych::corr.test(
    airquality[1:2], airquality[3:6], method = "spearman")
ggasycorr(psych_obj = cor_res)
# The positive correlation between solar radiation intensities and
# temperatures measured is not significant anymore.
#
# Do Spearman's correlation and apply Benjamini–Hochberg's correction
cor_res <- psych::corr.test(
    airquality[1:2], airquality[3:6], method = "spearman", adjust = "BH")
ggasycorr(psych_obj = cor_res)
# The positive correlation between solar radiation intensities and
# temperatures measured is significant again with Benjamini–Hochberg's
# procedure
#
# Apply no P-values correction procedure
cor_res <- psych::corr.test(
    airquality[1:2], airquality[3:6], method = "spearman", adjust = "none")
ggasycorr(psych_obj = cor_res)
# Set a P-value cut-off at 0.01 instead of 0.05 (default)
cor_res <- psych::corr.test(
    airquality[1:2], airquality[3:6], method = "spearman", adjust = "BH")
ggasycorr(psych_obj = cor_res, pval_cutoff = 0.01)
# The positive correlation between solar radiation intensities and
# temperatures measured is no longer significant at P-value <= 0.01
#
# You can also choose to display only results with a P-value below a maximum
# value:
ggasycorr(psych_obj = cor_res, show_pval_below = 0.1)
# And/or results above a minimum absolute correlation:
ggasycorr(psych_obj = cor_res, show_abs_cor_above = 0.5)
# Customize X & Y labels and add title using ggplot2 grammar
ggasycorr(psych_obj = cor_res, pval_cutoff = 0.01) +
    labs(
        x = "Weather & Time of the year", y = "Conditions",
        title = "Spearman's correlation of ozone vs. wind & temperature")
}
\author{
Yoann Pageaud.
}
