% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcoverage.R
\name{ggcoverage}
\alias{ggcoverage}
\title{Plots an annotated barplot.}
\usage{
ggcoverage(
  data,
  round.unit = 2,
  rev.stack = FALSE,
  invert.percent = FALSE,
  horizontal = FALSE,
  log.scaled = FALSE,
  decreasing.order = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing:
\itemize{
 \item{labels in column 1}
 \item{total amount in column 2}
 \item{subset amount in column 3}
}}

\item{round.unit}{An \code{integer} to specify how much decimals should be
kept when rounding percentages.}

\item{horizontal}{A \code{logical} to specify whether the plot should be
plotted horizontally or vertically.}
}
\value{
A \code{gg} dodged barplot with annotations.
}
\description{
Plots an annotated barplot.
}
\author{
Yoann Pageaud.
}
