% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{ggdend}
\alias{ggdend}
\title{Creates a dendogram in ggplot2.}
\usage{
ggdend(df, orientation)
}
\arguments{
\item{df}{A \code{data.frame} containing variables and values to be
used to create the dendrogram.}

\item{orientation}{A \code{character} specifying the orientation of the
dendrogram. Possible values are "top" and "left".}
}
\value{
A \code{gg} plot of the dendrogram.
}
\description{
Creates a dendogram in ggplot2.
}
\author{
Yoann Pageaud.
}
