% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{ggdend}
\alias{ggdend}
\title{Creates a dendogram in ggplot2.}
\usage{
ggdend(df, orientation, plot.type)
}
\arguments{
\item{df}{A \code{data.frame} containing variables and values to be
used to create the dendrogram.}

\item{orientation}{A \code{character} specifying the orientation of the
dendrogram. Possible values are "top" and "left".}

\item{plot.type}{A \code{character} specifying whether the plot is a
correlation plot ('corrplot') or a heatmap ('heatmap').
This parameter is used to know if a dendrogram displayed
on the left of a plot should be vertically reversed (like
in a correlation plot) or not (like in a heatmap).}
}
\value{
A \code{gg} plot of the dendrogram.
}
\description{
Creates a dendogram in ggplot2.
}
\author{
Yoann Pageaud.
}
