% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{ggdend}
\alias{ggdend}
\title{Creates a dendogram in ggplot2.}
\usage{
ggdend(df, orientation, reverse.x = FALSE, theme_dend = NULL)
}
\arguments{
\item{df}{A \code{data.frame} containing variables and values to be
used to create the dendrogram.}

\item{orientation}{A \code{character} specifying the orientation of the
dendrogram. Possible values are "top" and "left".}

\item{reverse.x}{A \code{logical} specifying whether the X-axis should be
reversed (reverse.x = TRUE like in a correlation plot), or
kept unchanged (reverse.x = TRUE like in a heatmap), for a
dendrogram to be displayed on the left of a plot
(Default: reverse.x = FALSE).}

\item{theme_dend}{A ggplot2 \code{theme} to specify any theme parameter you
wish to custom the dendrogram
(Default: theme_dend = NULL). For more information about
how to define a theme, see \link[ggplot2]{theme}.}
}
\value{
A \code{gg} plot of the dendrogram.
}
\description{
Creates a dendogram in ggplot2.
}
\examples{
# Create matrix from mtcars dataset
mat <- as.matrix(t(scale(mtcars)))
# Compute Pearson's correlation between cars
corr.res <- psych::corr.test(
    mat, use = "pairwise", method = "pearson", adjust = "BH")$r
# Normalize the correlation matrix
norm.res <- (corr.res-min(corr.res))/(max(corr.res)-min(corr.res))
# Do a complete hierarchy clustering on the normalized matrix
hierarchy.clust <- fastcluster::hclust(
    d = stats::as.dist(1-norm.res), method = "complete")
# Convert the cluster object into a dendrogram object
ddgr <- stats::as.dendrogram(hierarchy.clust)
ddgr_dat <- ggdendro::dendro_data(ddgr)
# Draw the dendrogram vertically
ggdend(df = ddgr_dat$segments, orientation = "top")
# Same horizontally
ggdend(df = ddgr_dat$segments, orientation = "left")
# Reverse branches order
ggdend(df = ddgr_dat$segments, orientation = "left", reverse.x = TRUE)
}
\author{
Yoann Pageaud.
}
