% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfusion.R
\name{ggfusion.corr}
\alias{ggfusion.corr}
\title{Draws 2 triangle matrices of computed pairwise correlations' results.}
\usage{
ggfusion.corr(
  data,
  upper.corr = "pearson",
  upper.value = "r",
  lower.corr = "pearson",
  lower.value = "p",
  na.rm = "pairwise",
  order.method = "hclust",
  order.select = "upper",
  hclust.method = "complete",
  p.adjust = "BH",
  dendrograms = FALSE,
  dend.size = 1,
  annot.grps = list(Groups = seq(ncol(data))),
  annot.pal = grDevices::rainbow(n = ncol(data)),
  annot.size = 1,
  upper_theme = NULL,
  upper_scale_fill_grad = ggplot2::scale_fill_gradientn(colors =
    BiocompR::biopalette(name = "RColorBrewer_RdBu8"), name = "Upper\\ntriangle values",
    limits = c(-1, 1)),
  upper_guide_custom_bar = ggplot2::guide_colorbar(barheight = 10, barwidth = 0.7,
    ticks.linewidth = 0.5, ticks.colour = "black", frame.linewidth = 0.5, frame.colour =
    "black"),
  lower_theme = NULL,
  lower_scale_fill_grad = ggplot2::scale_fill_gradient2(low = "darkblue", mid = "white",
    high = "darkred", midpoint = -log10(0.05), name = "Lower\\ntriangle values"),
  lower_guide_custom_bar = ggplot2::guide_colorbar(barheight = 0.7, barwidth = 10,
    ticks.linewidth = 0.5, ticks.colour = "black", frame.linewidth = 0.5, frame.colour =
    "black"),
  plot_title = NULL,
  grid_col = "white",
  grid_linewidth = 0.3,
  diagonal_col = "white",
  lgd.merge = FALSE,
  lgd.ncol = NULL,
  lgd.space.height = 26,
  lgd.space.width = 1,
  ncores = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{A \code{matrix} or \code{dataframe}.}

\item{upper.corr}{The comparison for which results will be
displayed in the upper triangle of the plot as a
\code{character} matching one of these:
'pearson','spearman','kendall'.}

\item{upper.value}{A \code{character} matching the type of value to
                            display in the upper triangle of the plot
                            resulting from the comparison.\cr The possible
                            value types are:
\itemize{
 \item{'r' - the correlation values.}
 \item{'n' - the number of cases used for the comparisons.}
 \item{'stat' - the values of the comparison statistic.}
 \item{'p' - the P-values of the comparison (-log10() transformed).}
 \item{'se' - the standard errors of the comparison.}
}}

\item{lower.corr}{The comparison for which results will be
displayed in the lower triangle of the plot as a
\code{character} matching one of these:
'pearson','spearman','kendall'.}

\item{lower.value}{See 'upper.value' help.}

\item{na.rm}{A \code{character} to specify how to handle
missing values when calculating a correlation.
Possible types are 'pairwise' and 'complete'.
'pairwise' is the default value and will do
pairwise deletion of cases. 'complete' will
select just complete cases.}

\item{order.method}{A \code{character} specifying the ordering method
to apply.
\cr Possible ordering methods are:
\itemize{
 \item{'AOE' - angular order of the eigenvectors.}
 \item{'FPC' - first principal component order.}
 \item{'hclust' - hierarchical clustering order.}
 \item{'alphabet' - alphabetical order}
 \item{'default'}
}}

\item{order.select}{A \code{character} specifying comparison results
to use for the clustering.\cr Possible values
are 'upper' or 'lower'.}

\item{hclust.method}{A \code{character} specifying the method to use
for the hierarchical clustering if the ordering
method is 'hclust'.\cr Possible methods are:
'ward.D','ward.D2', 'single', 'complete',
'average' (= UPGMA), 'mcquitty' (= WPGMA),
'median' (= WPGMC) or 'centroid' (= UPGMC).}

\item{p.adjust}{A \code{character} specifying what adjustment
for multiple tests should be used.\cr
(Default: p.adjust = "BH"; Supported:
p.adjust = c("holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none")).}

\item{annot.grps}{A \code{list} of vectors of groups to which
variables belongs for the annotation sidebars.
Vectors' lengths have to match the number of
variables.}

\item{annot.pal}{A \code{vector} or a list of vectors containing
colors as characters for the annotation
sidebars. The length of vectors has to match the
number of levels of vectors listed in
'annot.grps'. If a list is provided, its length
must match the length of the list provided to
'annot.grps'.}

\item{annot.size}{An \code{integer} to increase or decrease the
size of the annotation side bar.}

\item{grid_col}{A \code{character} specifying the color of the
grid.}

\item{grid_linewidth}{A \code{double} value for the thickness of the
grid.}

\item{diagonal_col}{A \code{character} defining the color of cells
with of the empty diagonal.}

\item{ncores}{An \code{integer} to specify the number of
cores/threads to be used to parallel-run tests.}

\item{dendro.size}{An \code{integer} to increase or decrease the
size of the dendrogram.}
}
\value{
A \code{gtable} object plotted automatically and a \code{list} of
       results from the 2 comparisons as 2 matrices, with the same gtable
       object:
       \itemize{
        \item{'upper.res' - the matrix result of pairwise comparisons
        displayed in the upper triangle.}
        \item{'lower.res' - the matrix result of pairwise comparisons
        displayed in the lower triangle.}
        \item{'fused.plot' - the plot.}
       }
}
\description{
Draws 2 triangle matrices of computed pairwise correlations' results.
}
\examples{
# Default fusion correlation plot showing Pearson correlations & P-values
mat <- as.matrix(t(scale(mtcars)))
res <- ggfusion.corr(data = mat)
# Same thing using the Spearman correlation
res <- ggfusion.corr(
    data = mat, upper.corr = "spearman", lower.corr = "spearman")
# Fuse correlations results from Pearson (top) and Spearman (bottom)
res <- ggfusion.corr(
    data = mat, upper.corr = "pearson", upper.value = "r",
    lower.corr = "spearman", lower.value = "r",
    lower_scale_fill_grad = ggplot2::scale_fill_gradientn(
        colors = BiocompR::biopalette(name = "RColorBrewer_RdBu8"),
        limits = c(-1, 1))) # Setting the scale of the Spearman correlations
# Set the missing data removal method to complete before computing
# correlations
res <- ggfusion.corr(data = mat, na.rm = "complete")
# Order samples following the angular order of eigenvectors
# (of the 2 first eigenvalues on the matrix selected).
res <- ggfusion.corr(data = mat, order.method = "AOE")
# Select the lower matrix (instead of the upper) to apply the order method
res <- ggfusion.corr(
    data = mat, order.method = "AOE", order.select = "lower")
}
\references{
\href{https://www.scholars.northwestern.edu/en/publications/psych-procedures-for-personality-and-psychological-research}{William R Revelle, psych: Procedures for Personality and Psychological Research. Northwestern University, Evanston, Illinois, USA (2017).}

\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6099145/#B13}{Why, When and How to Adjust Your P Values? Mohieddin Jafari and Naser Ansari-Pour - Cell J. 2019 Winter; 20(4): 604–607.}
}
\author{
Yoann Pageaud.
}
