% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_plot.R
\name{ggfusion.free}
\alias{ggfusion.free}
\title{Displays 2 triangle matrices fused together in a single plot.}
\usage{
ggfusion.free(
  sample.names,
  upper.mat,
  lower.mat,
  order.select,
  order.method,
  hclust.method,
  annot.grps = list(Groups = seq(length(sample.names))),
  annot.pal = grDevices::rainbow(n = length(sample.names)),
  annot.size = 1,
  lgd.merge = FALSE,
  lgd.ncol = NULL,
  lgd.space.height = 26,
  lgd.space.width = 1,
  dendro.pos = "none",
  dendro.size = 0,
  grid_col = "white",
  grid_linewidth = 0.3,
  upper_theme = NULL,
  upper_scale_fill_grad = ggplot2::scale_fill_gradientn(colors =
    BiocompR::biopalette(name = "viridis_C_plasma"), name = "Upper\\ntriangle values"),
  upper_guide_custom_bar = ggplot2::guide_colorbar(barheight = 10, barwidth = 0.7,
    ticks.linewidth = 0.5, ticks.colour = "black", frame.linewidth = 0.5, frame.colour =
    "black"),
  lower_theme = NULL,
  lower_scale_fill_grad = ggplot2::scale_fill_gradientn(colors =
    BiocompR::biopalette(name = "viridis_D_viridis"), name = "Lower\\ntriangle values"),
  lower_guide_custom_bar = ggplot2::guide_colorbar(barheight = 0.7, barwidth = 10,
    ticks.linewidth = 0.5, ticks.colour = "black", frame.linewidth = 0.5, frame.colour =
    "black"),
  diagonal_col = "white",
  plot_title = NULL,
  verbose = FALSE
)
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Displays 2 triangle matrices fused together in a single plot.
}
\author{
Yoann Pageaud.
}
