% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{ggtriangle}
\alias{ggtriangle}
\title{Draws a triangle ggplot for a basic upper or lower triangle matrix.}
\usage{
ggtriangle(
  melt_tri,
  grid_col = "white",
  grid_linewidth = 0.3,
  ggtri_theme = NULL,
  scale_fill_grad = ggplot2::scale_fill_gradientn(colors = BiocompR::biopalette(name =
    "viridis_C_plasma"), na.value = "grey"),
  guide_custom_bar = ggplot2::guide_colorbar(ticks.linewidth = 0.5, ticks.colour =
    "black", frame.colour = "black", frame.linewidth = 0.5)
)
}
\arguments{
\item{melt_tri}{A \code{data.frame} melted triangle containing a
statistical test values.}

\item{grid_col}{A \code{character} specifying the color of the grid.}

\item{grid_linewidth}{A \code{double} value for the thickness of the grid.}

\item{ggtri_theme}{A ggplot2 \code{theme} to specify any theme parameter
you wish to custom on the triangle plot
(Default: ggtri_theme = NULL). For more information
about how to define a theme, see
\link[ggplot2]{theme}.}

\item{scale_fill_grad}{A \code{ScaleContinous} object generated by ggplot2
functions such as \link[ggplot2]{scale_fill_gradient}
,\link[ggplot2]{scale_fill_gradient2} or
\link[ggplot2]{scale_fill_gradientn} to customize
heatmap colors and the associated color bar.}

\item{guide_custom_bar}{A \code{guide} object generated by the ggplot2
function \link[ggplot2]{guide_colorbar} to custom the
triangle plot color bar appearance
(see also 'scale_fill_grad' option).}
}
\value{
A \code{gg} object of a basic triangle plot (a 'geom_tile()').
}
\description{
Draws a triangle ggplot for a basic upper or lower triangle matrix.
}
\author{
Yoann Pageaud.
}
