% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{ggvolcano.corr}
\alias{ggvolcano.corr}
\title{Plots results of correlation test between a single variable and multiple
others as volcano plot.}
\usage{
ggvolcano.corr(
  data,
  p.cutoff = 0.01,
  corr.cutoff = NULL,
  title.corr.cutoff = "Correlation cut-off",
  corr.label.cutoff
)
}
\arguments{
\item{data}{A \code{data.table} with 3 to 5 columns:
\itemize{
 \item{column 1 - strings to be used as labels for
 individual dots.}
 \item{column 2 - correlation values.}
 \item{column 3 - correlation p-values.}
 \item{column 4 (optional) - groups to be used for
 coloring the dots.}
 \item{column 5 (optional) - values to be used to
 define dots sizes. It can be the sample size used
 for the calculation of the correlation test.}
}}

\item{p.cutoff}{A \code{numeric} between 0 and 1 to be used as a
maximum cut-off on p-values
(Default: p.cutoff = 0.01).}

\item{corr.cutoff}{A \code{numeric} value between -1 and 1 to be used
as a cut-off on correlation values.
\itemize{
 \item{if corr.cutoff < 0: it will be a maximum
 cut-off on negative correlation values.}
 \item{if corr.cutoff > 0: it will be a minimum
 cut-off on positive correlation values.}
}
It is not used as an absolute cut-off on positive
and negative correlation values as if the sign of
the correlation value is different from the one of
the cut-off, it might not necessary make sense apply
the cut-off to it.}

\item{title.corr.cutoff}{A \code{character} to be used as the label of the
correlation cut-off vertical blue line
(Default: title.corr.cutoff="Correlation cut-off").}

\item{corr.label.cutoff}{A \code{numeric} vector of 1 or 2 values, between -1
and 1 to be used as the minimum cut-off on
positive and negative correlation values.
\itemize{
 \item{If 1 value is given, it will be define as the
 minimum cut-off on absolute correlation values
 (positive and negative ones).}
 \item{If 2 values are given, the smalest one will
 be used as a maximum cut-off on negative
 correlation values. the biggest one will be used as
 a minimum cut-off on positive correlation values.
 The smalest value must be comprised between -1 and
 0. The biggest value must be comprised between 0
 and 1.}
}}
}
\value{
A \code{gg} volcano plot object.
}
\description{
Plots results of correlation test between a single variable and multiple
others as volcano plot.
}
\author{
Yoann Pageaud.
}
