% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{ggvolcano.corr}
\alias{ggvolcano.corr}
\title{Plots results of correlation test between a single variable and multiple
others as volcano plot.}
\usage{
ggvolcano.corr(
  data,
  p.cutoff = 0.01,
  corr.cutoff = NULL,
  title.cutoff = "Correlation cut-off",
  label.cutoff = NULL,
  x.col.sign = FALSE,
  force.label = NULL
)
}
\arguments{
\item{data}{A \code{data.table} with 3 to 5 columns:
\itemize{
 \item{column 1 - The labels.}
 \item{column 2 - correlation values.}
 \item{column 3 - P-values.}
 \item{column 4 (optional) - groups to be used to color
 points. If none provided, you can use 'x.col.sign' to
 color dots based on the correlation category they belong
 to.}
 \item{column 5 (optional) - values to be used to define
 dots sizes based on another variable.}
}}

\item{p.cutoff}{A \code{numeric} between 0 and 1 to be used as a maximum
cut-off on p-values (Default: p.cutoff = 0.01).}

\item{corr.cutoff}{A \code{numeric} vector of 1 or 2 values, comprised
between -1 and 1, to be used as a cut-off on correlation
values (Default: corr.cutoff = NULL).
\itemize{
 \item{If 1 value is given, it will be defined as the
 minimum cut-off on absolute correlation values
 (positive and negative ones).}
 \item{If 2 values are given: The smallest one will be
 used as a maximum cut-off on negative correlation
 values. The biggest one will be used as a minimum
 cut-off on positive correlation values.
 The smallest value must be inferior or equal to 0. The
 biggest value must be superior or equal to 0.}
}
corr.cutoff value(s) will be used to draw vertical lines
on the volcano plot.}

\item{title.cutoff}{A \code{character} to specify how you wish to label the
correlation limits on the volcano plot
(Default: title.cutoff = "Correlation cut-off").}

\item{label.cutoff}{A \code{numeric} vector of 1 or 2 values, comprised
between -1 and 1, to be used as a cut-off on label
display
(Default: label.cutoff = NULL does not show any label).
\itemize{
 \item{If 1 value is given, it will be defined as the
 minimum cut-off on absolute correlation values
 (positive and negative ones).}
 \item{If 2 values are given: The smallest one will be
 used as a maximum cut-off on negative correlation
 values. The biggest one will be used as a minimum
 cut-off on positive correlation values.
 The smallest value must be inferior or equal to 0. The
 biggest value must be superior or equal to 0.}
}
Dots outside these limits will be labeled. Dots within
the range of these limits will remain unlabeled.}

\item{x.col.sign}{A \code{logical} to activate automatic coloring of data
based on their correlation category. If TRUE, elements
will be divided into 3 categories:
\itemize{
 \item{Blue - elements below the negative 'corr.cutoff'.}
 \item{Red - elements above the positive 'corr.cutoff'.}
 \item{Grey - any other element not meeting the
 requirements in the first nor the second categories.}
}
If FALSE, dots will be colored using groups from data
(if any provided). If no groups are passed within data,
dots will remain black.}

\item{force.label}{A \code{character} vector matching elements in the
first column of 'data' that you wish to label,
independently from their actual significance
(Default: force.label = NULL). force.label overrides
other labeling parameters.}
}
\value{
A \code{gg} volcano plot of your correlation test results.
}
\description{
Plots results of correlation test between a single variable and multiple
others as volcano plot.
}
\examples{
# Calculate correlation on mtcars dataset
corr_res <- Hmisc::rcorr(x = as.matrix(mtcars))
# Create data.table with labels, correlation values, and P values
dt_corr <- data.table(
  labels = rownames(corr_res$r)[-nrow(corr_res$r)],
  correlation = corr_res$r[-nrow(corr_res$r), "carb"],
  p_value = corr_res$P[-nrow(corr_res$r), "carb"])
# Draw basic correlation volcano plot
ggvolcano.corr(data = dt_corr)
# Suppress the alpha warning when drawing a volcano
warn.handle(
  print(ggvolcano.corr(data = dt_corr)),
  pattern = "Using alpha for a discrete variable is not advised.")
# Set another P-value cutoff
ggvolcano.corr(data = dt_corr, p.cutoff = 0.05)
# Set an absolute correlation cutoff
ggvolcano.corr(data = dt_corr, corr.cutoff = 0.4)
# Set different negative and positive corelation cutoffs
ggvolcano.corr(data = dt_corr, corr.cutoff = c(-0.3, 0.5))
# Rename correlation cutoff title
ggvolcano.corr(
  data = dt_corr, corr.cutoff = 0.5,
  title.cutoff = "Minimum absolute correlation")
# Set a negative and positive correlation cutoff to show/hide point labels
ggvolcano.corr(data = dt_corr, corr.cutoff = 0.4, label.cutoff = c(-0.4, 1))
# Turn on autocolor based on X-axis sign of values and significance
ggvolcano.corr(data = dt_corr, corr.cutoff = c(-0.6, 0.5), x.col.sign = TRUE)
# Force a specific label display on not significantly correlated variable
ggvolcano.corr(data = dt_corr, force.label = "am")
# Force display of all point labels significant or not
ggvolcano.corr(
  data = dt_corr, force.label = rownames(corr_res$r)[-nrow(corr_res$r)])
# Add 'group' data for colors
dt_corr[, var_group := c(rep("1", 5), rep("2", 5))]
ggvolcano.corr(data = dt_corr)
# Add 'importance' data for point sizes
dt_corr[, influence := c(rep(1, 3), rep(2, 3), rep(3, 4))]
ggvolcano.corr(data = dt_corr)
# Custom the volcano with ggplot2 components
ggvolcano.corr(data = dt_corr) +
  ggtitle("Correlation of mtcars variables with carburators") +
  theme(plot.title = element_text(size = 14)) +
  scale_color_manual(values = c("purple", "orange")) +
  scale_size(breaks = c(1, 2, 3))
}
\author{
Yoann Pageaud.
}
