% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{ggvolcano.test}
\alias{ggvolcano.test}
\title{Plots results of statistical tests as volcano plot.}
\usage{
ggvolcano.test(
  data,
  p.cutoff = 0.01,
  l2fc.cutoff = NULL,
  title.fc.cutoff = "L2FC cut-off",
  label.cutoff = 0
)
}
\arguments{
\item{data}{A \code{data.table} with 3 to 5 columns:
\itemize{
 \item{column 1 - The labels.}
 \item{column 2 - log2(fold changes) values.}
 \item{column 3 - P-values.}
 \item{
 column 4 (optional) - groups to be used to color points.
 If none provided, elements will be divided into
 3 categories:
 \itemize{
  \item{Red - elements above the positive
  'l2fc.cutoff' and below 'p.cutoff'.}
  \item{Blue - elements below the negative
  'l2fc.cutoff' and below 'p.cutoff'.}
  \item{Black - any other element not meeting the
  requirements in the first nor the second category.}
 }}
}}

\item{p.cutoff}{A \code{numeric} between 0 and 1 to be used as a maximum
cut-off on p-values (Default: p.cutoff = 0.01).}

\item{l2fc.cutoff}{A \code{numeric} to be used as a cut-off on fold changes
(Default: l2fc.cutoff = 0).}

\item{label.cutoff}{A \code{numeric} vector of 1 or 2 values, to be used as
the minimum cut-off on positive and negative correlation
values.
\itemize{
 \item{If 1 value is given, it will be define as the
 minimum cut-off on absolute correlation values (positive
 and negative ones).}
 \item{If 2 values are given, the smalest one will be
 used as a maximum cut-off on negative log2(fold change)
 values. The biggest one will be used as a minimum
 cut-off on positive log2(fold change) values.
 The smalest value must be inferior to 0. The biggest
 value must be superior to 0.}
}}

\item{label}{A \code{character} vector to be used to define which
elements in the \code{data.table} column 1 should be
labeled on the plot. (Default: label = NULL).}

\item{group.label}{A \code{logical} to be used to define whether labeling is
done on individual (Default: group.label = FALSE) or group
level (group.label = TRUE). The parameter is only used
when the \code{data.table} column 4 is set.}
}
\value{
A \code{gg} volcano plot.
}
\description{
Plots results of statistical tests as volcano plot.
}
\author{
Verena Bitto, Yoann Pageaud.
}
