% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{ggvolcano.test}
\alias{ggvolcano.test}
\title{Plots results of statistical tests as a volcano plot.}
\usage{
ggvolcano.test(
  data,
  p.cutoff = 0.01,
  l2fc.cutoff = NULL,
  title.cutoff = "L2FC cut-off",
  label.cutoff = NULL,
  x.col.sign = FALSE,
  l2.transform = FALSE,
  force.label = NULL
)
}
\arguments{
\item{data}{A \code{data.table} with 3 to 5 columns:
\itemize{
 \item{column 1 - The labels.}
 \item{column 2 - log2(fold changes) values.}
 \item{column 3 - P-values.}
 \item{column 4 (optional) - groups to be used to color
 points. If none provided, you can use 'x.col.sign' to
 color dots based on the fold change category they belong
 to.}
 \item{column 5 (optional) - values to be used to define
 dots sizes based on another variable.}
}}

\item{p.cutoff}{A \code{numeric} between 0 and 1 to be used as a maximum
cut-off on p-values (Default: p.cutoff = 0.01).}

\item{l2fc.cutoff}{A \code{numeric} vector of 1 or 2 values, to be used as
a cut-off on log2(fold change) values
(Default: l2fc.cutoff = NULL).
\itemize{
 \item{If 1 negative value is given, it will be defined
 as the maximum cut-off on negative log2(fold change)
 values.}
 \item{If 1 positive value is given, it will be defined
 as the minimum cut-off on positive log2(fold change)
 values.}
 \item{If 2 values are given: The smallest one will be
 used as a maximum cut-off on negative log2(fold change)
 values. The biggest one will be used as a minimum
 cut-off on positive log2(fold change) values.
 The smallest value must be inferior or equal to 0. The
 biggest value must be superior or equal to 0.}
}
l2fc.cutoff value(s) will be used to draw vertical lines
on the volcano plot.}

\item{title.cutoff}{A \code{character} to specify how you wish to label the
log2(fold change) limits on the volcano plot
(Default: title.cutoff = "L2FC cut-off").}

\item{label.cutoff}{A \code{numeric} vector of 1 or 2 values, to be used as
a cut-off on log2(fold change) values
(Default: label.cutoff = NULL does not show any label).
\itemize{
 \item{If 1 value is given, it will be defined as the
 minimum cut-off on absolute log2(fold change) values
 (positive and negative ones).}
 \item{If 2 values are given: The smallest one will be
 used as a maximum cut-off on negative log2(fold change)
 values. The biggest one will be used as a minimum
 cut-off on positive log2(fold change) values.
 The smallest value must be inferior or equal to 0. The
 biggest value must be superior or equal to 0.}
}
Dots outside these limits will be labeled. Dots within
the range of these limits will remain unlabeled.}

\item{x.col.sign}{A \code{logical} to activate automatic coloring of data
based on their fold change category. If TRUE, elements
will be divided into 3 categories:
\itemize{
 \item{Blue - elements below the negative 'l2fc.cutoff'.}
 \item{Red - elements above the positive 'l2fc.cutoff'.}
 \item{Grey - any other element not meeting the
 requirements in the first nor the second categories.}
}
If FALSE, dots will be colored using groups from data
(if any provided). If no groups are passed within data,
dots will remain black.}

\item{l2.transform}{A \code{logical} to specify whether the data in the
second column should be log2 transformed
(l2.transform = TRUE) or if the log2 transformation has
already been applied and no further transformation is
needed (Default: l2.transform = FALSE).}

\item{force.label}{A \code{character} vector matching elements in the
first column of 'data' that you wish to label,
independently from their actual significance
(Default: force.label = NULL). force.label overrides
other labeling parameters.}
}
\value{
A \code{gg} volcano plot of your statistical test results.
}
\description{
Plots results of statistical tests as a volcano plot.
}
\examples{
# Compute t.test and fold change of Tooth Growth in Guinea Pigs
dt <- as.data.table(ToothGrowth)
ls_res <- lapply(unique(dt$dose), FUN = function(d){
    log2fc <- mean(log2(dt[dose == d & supp == "VC"]$len)) - mean(
        log2(dt[dose == d & supp == "OJ"]$len))
    p.val <- t.test(
        x = dt[dose == d & supp == "VC"]$len,
        y = dt[dose == d & supp == "OJ"]$len)$p.value
    data.table("dose" = as.character(d), "logfc" = log2fc, "p.val" = p.val)
})
dt_res <- rbindlist(ls_res)
# Most basic volcano plot on statistical test results
ggvolcano.test(data = dt_res)
# Set another P-value cutoff
ggvolcano.test(data = dt_res, p.cutoff = 0.05)
# Set an absolute log2 fold change cutoff
ggvolcano.test(data = dt_res, l2fc.cutoff = c(-0.4, 0.4))
# Set different negative and positive log2 fold change cutoffs
ggvolcano.test(data = dt_res, l2fc.cutoff = c(-0.3, 0.5))
# Rename log2 fold change cutoff title
ggvolcano.test(
  data = dt_res, l2fc.cutoff = c(-0.4, 0.4),
  title.cutoff = "Minimum absolute log2(FC)")
# Set a negative and positive log2 fold change cutoff to show/hide labels
ggvolcano.test(
  data = dt_res, l2fc.cutoff = c(-0.4, 0.4), label.cutoff = c(-0.4, 0))
# Turn on autocolor based on X-axis sign of values and significance
ggvolcano.test(data = dt_res, l2fc.cutoff = c(-0.4, 0), x.col.sign = TRUE)
}
\author{
Yoann Pageaud, Verena Bitto.
}
