% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{heatmap_layout}
\alias{heatmap_layout}
\title{Draft using the egg package for the gg2heatmap layout}
\usage{
heatmap_layout()
}
\arguments{
\item{param1}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Draft using the egg package for the gg2heatmap layout
}
\author{
Yoann Pageaud.
}
\keyword{internal}
