% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{heatmap_layout}
\alias{heatmap_layout}
\title{Builds the gg2heatmap layout using the egg package}
\usage{
heatmap_layout(
  dd.rows = FALSE,
  dd.cols = FALSE,
  show.annot = FALSE,
  ddgr_seg_row = NULL,
  ddgr_seg_col = NULL,
  sidebar,
  htmp,
  legends,
  dend.row.size = 0,
  dend.col.size = 0,
  annot.size = 0,
  lgd.layout,
  lgd.space.width,
  override_grob_list = NULL,
  override_grob_height = NULL
)
}
\value{
A \code{egg} object of the assembled heatmap.
}
\description{
Builds the gg2heatmap layout using the egg package
}
\author{
Yoann Pageaud.
}
\keyword{internal}
