% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_na.R
\name{manage.na}
\alias{manage.na}
\title{Keeps, removes or imputes missing values in a matrix or a data.frame
accounting for group size.}
\usage{
manage.na(
  data,
  method = "remove",
  groups,
  grp.fun = "median",
  impute.fun = "median",
  vinterpolate = FALSE,
  row.na.cut = NULL,
  ncores = 1
)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} with column names.}

\item{method}{A \code{character} which defines the method to use for
managing NAs:
\itemize{
 \item{'keep' does nothing.}
 \item{'remove' removes all rows containing at least 1
 NA.}
 \item{'impute' removes rows containing only NAs and
 impute missing values in remaining rows. The imputation
 uses the groups defined in the parameter 'groups': e.g.
 in row 1 if the value of sample_1 is NA, the missing
 value will be imputed using the median of values from
 other samples of the same group than sample_1.}
}}

\item{groups}{A vector of length ncol(data) specifying the groups to
which samples belong. Order of groups in the vector has
to match the order of column names in data to properly
associate samples and groups.}

\item{impute.fun}{A \code{character} specifying the function to be used for
imputation (Default: impute.fun = "median").}

\item{vinterpolate}{A \code{logical} specifying whether vertical
interpolation should be applied on missing values, based
on previous and next available values in a column
(vinterpolate = TRUE) or not
(Default: vinterpolate = FALSE). It is combined with the
function use for imputation by group. vinterpolate is not
use if method is not set to 'impute'.}

\item{row.na.cut}{An \code{integer} specifying the cut-off for imputation
on rows missing values. if NULL, all rows with missing
values will be processed. if row.na.cut equals an
integer, only rows containing a number of NAs smaller or
equal to the cut-off will be processed
(Default: row.na.cut = NULL).}

\item{ncores}{An \code{integer} to specify the number of cores/threads
to be used to parallel-process the matrix
(Default: ncores = 1).}
}
\value{
A \code{matrix}.
}
\description{
Keeps, removes or imputes missing values in a matrix or a data.frame
accounting for group size.
}
\examples{
# Creating a data.frame of 4 columns with 50 values in each:
df <- data.frame("col1" = rnorm(n = 50, mean = 25, sd = 4),
                 "col2" = rnorm(n = 50, mean = 25, sd = 4),
                 "col3" = rnorm(n = 50, mean = 25, sd = 4),
                 "col4" = rnorm(n = 50, mean = 25, sd = 4))
# col1 and col4 contain missing data
df$col1[1:10] <- NA
df$col4[40:50] <- NA
df
# Create 2 groups: grp1 & grp2.
# col1 and col2 are in grp1, col3 and col4 are in grp2.
grps <- c("grp1","grp1","grp2","grp2")
# Remove all rows containing at least 1 missing value
manage.na(data = df, method = "remove", groups = grps)
# Impute by group the missing data (and remove rows missing all values)
manage.na(data = df, method = "impute", groups = grps)
}
\author{
Yoann Pageaud.
}
