% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{map.cat2pal}
\alias{map.cat2pal}
\title{Maps annotations categories to palettes.}
\usage{
map.cat2pal(origin.grps, groups, annot.pal)
}
\arguments{
\item{origin.grps}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Maps annotations categories to palettes.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
