% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_stat_tests.R
\name{pairwise.ks}
\alias{pairwise.ks}
\title{Computes a Kolmogrov-Smirnov test between all columns of a data.frame.}
\usage{
pairwise.ks(data, statistic, ncores)
}
\arguments{
\item{data}{A \code{data.frame} of numerical values.}

\item{statistic}{A \code{character} specifying the type of statistic to
retrieve from the test\cr
(Supported: statistic = c('n','stat','p')).}

\item{ncores}{An \code{integer} to specify the number of cores/threads to
be used to parallel-run tests.}
}
\value{
A \code{list} containing a matrix of the statistic values, a
        data.frame of the pairwise KS raw test results, and a table of
        pairwise combinations.
}
\description{
Computes a Kolmogrov-Smirnov test between all columns of a data.frame.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
