% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{plot.col.sidebar}
\alias{plot.col.sidebar}
\title{Creates a colored side annotation bars in ggplot2.}
\usage{
\method{plot}{col.sidebar}(
  sample.names,
  annot.grps,
  annot.pal,
  annot.pos,
  cor.order,
  split.annot = FALSE,
  merge.lgd = FALSE,
  right = FALSE,
  lgd.lab = "Legends",
  lgd.title = NULL,
  axis.text.x = element_text(size = 12),
  axis.text.y = element_text(size = 12),
  axis.ticks.x,
  axis.ticks.y,
  axis.title.x,
  axis.title.y,
  set.x.title,
  set.y.title,
  dendro.pos
)
}
\arguments{
\item{sample.names}{A \code{character} vector of the labels to be used for
annotation.}

\item{annot.grps}{A \code{list} of vectors of groups to which variables
belongs for the annotation sidebars. vectors lengths
have to match the number of variables.}

\item{annot.pal}{A \code{vector} or a list of vectors containing colors
as characters for the annotation sidebars. The length of
vectors has to match the number of levels of vectors
listed in 'annot.grps'.
\itemize{
 \item{If annot.pal is a list: its length must match the
       length of the list provided to 'annot.grps'.}
 \item{If annot.pal is a vector: make sure that the
       levels content of annotations listed in
       'annot.grps' is the same, and that no annotation
       contains less or more levels than another one in
       'annot.grps'.}
}}

\item{annot.pos}{A \code{character} specifying the position of the
annotation sidebar.\cr Possible values are: 'top',
'left' or 'both'.}

\item{cor.order}{An \code{integer} vector to be used to reorder the labels
in 'sample.names'.}

\item{split.annot}{A \code{logical} to specify whether a separating space
should be inserted between annotation bars or not
(Default: split.annot = FALSE).}

\item{merge.lgd}{A \code{logical} to specify whether annotation legends
should be merged (annot.lgd.merge = TRUE) or remain
separated (annot.lgd.merge = FALSE)
(Default: annot.lgd.merge = FALSE).}

\item{right}{A \code{logical} to specify that Y-axis should be
displayed on the right side of the plot
(Default: right = FALSE).}

\item{lgd.lab}{A \code{character} to specify a title to the legend of
the plot, only if 'merge.lgd' = TRUE
(Default: lgd.lab = "Legends").}

\item{lgd.title}{An \code{element_text} object to setup legend titles
(Default: lgd.title = NULL).}

\item{axis.text.x}{An \code{element_text} object to setup X axis text
(Default: axis.text.x = element_text(size = 12)).}

\item{axis.text.y}{An \code{element_text} object to setup Y axis text
(Default: axis.text.y = element_text(size = 12)).}

\item{axis.ticks.x}{An \code{element_line} object to setup X axis ticks.}

\item{axis.ticks.y}{An \code{element_line} object to setup Y axis ticks.}

\item{axis.title.x}{An \code{element_text} object to setup X axis title.
\itemize{Exceptions:
 \item{If annot.pos == 'left', then axis.title.x =
 element_blank().}
 \item{If annot.pos == dendro.pos == 'top', then
 axis.title.x = element_blank().}
}}

\item{axis.title.y}{An \code{element_text} object to setup Y axis title.
\itemize{Exceptions:
 \item{If annot.pos == 'top', then axis.title.y =
 element_blank().}
 \item{If annot.pos == dendro.pos == 'left', then
 axis.title.y = element_blank().}
}}

\item{set.x.title}{A \code{character}to be used as the title for the X axis.}

\item{set.y.title}{A \code{character}to be used as the title for the Y axis.}

\item{dendro.pos}{A \code{character} specifying the position of the
dendrogram (Supported: dendro.pos = c('top','left')).}
}
\value{
A \code{list} of length 2:
        \itemize{
         \item{'sidebar' contains the colored sidebar plot.}
         \item{'legends' contains a list of legends as \code{gg} objects.}
        }
}
\description{
Creates a colored side annotation bars in ggplot2.
}
\author{
Yoann Pageaud.
}
