% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ggplot_fun.R
\name{plot.col.sidebar}
\alias{plot.col.sidebar}
\title{Creates a colored side annotation bars in ggplot2.}
\usage{
plot.col.sidebar(
  sample.names,
  annot.grps,
  annot.pal,
  annot.pos,
  cor.order,
  split.annot = TRUE,
  merge.lgd = FALSE,
  right = FALSE,
  lgd.title = "Legends",
  axis.text.x,
  axis.text.y,
  axis.ticks.x,
  axis.ticks.y,
  axis.title.x,
  axis.title.y,
  set.x.title,
  set.y.title,
  dendro.pos
)
}
\arguments{
\item{sample.names}{A \code{type} parameter description.}
}
\value{
A \code{type} object returned description.
}
\description{
Creates a colored side annotation bars in ggplot2.
}
\author{
Yoann Pageaud.
}
