% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_na.R
\name{row.impute.na}
\alias{row.impute.na}
\title{Imputes missing data in a matrix row.}
\usage{
row.impute.na(
  data,
  miss.mat = NULL,
  grp_tbl,
  nr = NULL,
  r,
  grp.fun = "median",
  impute.fun = "median",
  vinterpolate = FALSE
)
}
\arguments{
\item{data}{A \code{matrix}.}

\item{miss.mat}{A logical \code{matrix} where cells are TRUE when there
is a numeric value in data, or FALSE when there are NAs
in data. miss.mat is only used when vinterpolate is TRUE
(Default: miss.mat = NULL).}

\item{grp_tbl}{A \code{dataframe} with 2 columns:
\itemize{
 \item{samples - which contains all column names from
 data.}
 \item{groups - which contains the groups to which each
 column belong to.}
}}

\item{nr}{An \code{integer} specifying the total number of rows in
the matrix data.}

\item{r}{An \code{integer} specifying the position of the row of
interest for imputation in the matrix data.}

\item{impute.fun}{A \code{character} specifying the function to be used for
imputation (Default: impute.fun = "median").}

\item{vinterpolate}{A \code{logical} specifying whether vertical
interpolation should be applied on missing values, based
on previous and next available values in a column
(vinterpolate = TRUE) or not
(Default: vinterpolate = FALSE). It is combined with the
function use for imputation by group. vinterpolate is not
use if method is not set to 'impute'.}
}
\value{
A \code{numeric} vector containing the imputed values of the row for
        each groups of columns.
}
\description{
Imputes missing data in a matrix row.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
